% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_add_column.R
\name{add_column}
\alias{add_column}
\title{add_column}
\usage{
add_column(
  .data,
  ...,
  .before = NULL,
  .after = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{.data}{mass_data class}

\item{...}{dynamic-dots Name-value pairs, passed on to tibble().
All values must have the same size of .data or size 1.}

\item{.before}{One-based column index or column name where to add
the new columns, default: after last column.}

\item{.after}{One-based column index or column name where to add
the new columns, default: after last column.}

\item{.name_repair}{Treatment of problematic column names:
"minimal": No name repair or checks, beyond basic existence,
"unique": Make sure names are unique and not empty,
"check_unique": (default value), no name repair, but check they are unique,
"universal": Make the names unique and syntactic
a function: apply custom name repair (e.g., .name_repair = make.names for
names in the style of base R).
A purrr-style anonymous function, see rlang::as_function()
This argument is passed on as repair to vctrs::vec_as_names().
See there for more details on these terms and the strategies
used to enforce them.}
}
\value{
A mass_dataset class object
}
\description{
add_column. This function is from tibble.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
