% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjust_confounder.R
\name{adjust_confounder}
\alias{adjust_confounder}
\title{Adjust Confounders in a mass_dataset Object}
\usage{
adjust_confounder(object, confounder_name_list)
}
\arguments{
\item{object}{A \code{mass_dataset} object that contains expression data and sample information.}

\item{confounder_name_list}{A character vector specifying the names of the confounders
to adjust for. These names should be columns in the \code{sample_info} slot of the \code{mass_dataset} object.}
}
\value{
A modified \code{mass_dataset} object with adjusted expression data where confounders
have been accounted for.
}
\description{
This function adjusts for confounders in the expression data of a \code{mass_dataset} object.
The function takes a list of confounders from the \code{sample_info} data and performs a
linear regression adjustment on the expression data for each specified confounder.
}
\details{
This function performs a linear model regression for each feature in the expression data
against the specified confounders from the sample information. The residuals of the model
are used as the adjusted expression data. If any NA values are present in the confounders
or if the confounder names are not found in the \code{sample_info}, the function will stop
and return an error.
}
\section{Errors}{

\itemize{
\item The function will stop if \code{confounder_name_list} is empty.
\item The function will stop if any of the confounder names in \code{confounder_name_list}
are not present in the \code{sample_info}.
\item The function will stop if any NA values are present in the specified confounders.
}
}

\examples{
data("expression_data")
data("sample_info")
data("variable_info")

sample_info$batch <- rep(c(1, 2), length.out = nrow(sample_info))
sample_info$age <- seq_len(nrow(sample_info)) + 20

object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

adjusted_object <- adjust_confounder(
  object = object,
  confounder_name_list = c("batch", "age")
)

head(extract_expression_data(adjusted_object))


}
