% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R
\name{annotation_table}
\alias{annotation_table}
\title{Generic Function for Retrieving Annotation Table}
\usage{
annotation_table(object, ...)
}
\arguments{
\item{object}{An object from which the annotation table will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a dataframe or matrix containing annotation information for the object.
}
\description{
The \code{annotation_table} generic function retrieves the annotation table from the provided object.
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
