% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-mass_dataset.R
\name{$.mass_dataset}
\alias{$.mass_dataset}
\title{$ method}
\usage{
\method{$}{mass_dataset}(x, name)
}
\arguments{
\item{x}{A \code{mass_dataset} object.}

\item{name}{A column name from \code{expression_data}.}
}
\value{
The requested column from \code{expression_data}, if it exists.
}
\description{
Extracts a column from the \code{expression_data} slot of a
\code{mass_dataset} object by name.
}
