% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{cbind_mass_dataset}
\alias{cbind_mass_dataset}
\title{cbind mass_data class}
\usage{
cbind_mass_dataset(x, y, deparse.level = 1)
}
\arguments{
\item{x}{A \code{mass_dataset} object.}

\item{y}{Another \code{mass_dataset} object to be column bound with \code{x}.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. Default is 1.}
}
\value{
A \code{mass_dataset} object containing the combined samples and merged sample metadata.
}
\description{
cbind mass_data class
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
x <- object[, 1:2]
temp_expression_data <- extract_expression_data(x)
colnames(temp_expression_data) <- paste0(colnames(temp_expression_data), "_2")
temp_sample_info <- extract_sample_info(x)
temp_sample_info$sample_id <- colnames(temp_expression_data)
rownames(temp_sample_info) <- temp_sample_info$sample_id
y <- create_mass_dataset(
  expression_data = temp_expression_data,
  sample_info = temp_sample_info,
  variable_info = extract_variable_info(x),
  sample_info_note = extract_sample_info_note(x),
  variable_info_note = extract_variable_info_note(x)
)

result <- cbind_mass_dataset(x, y)
dim(result)

}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
