% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_annotation_table}
\alias{check_annotation_table}
\title{Validate Annotation Table Format}
\usage{
check_annotation_table(annotation_table)
}
\arguments{
\item{annotation_table}{A data frame containing the annotation data to be validated.}
}
\value{
The function returns a message indicating that the \code{annotation_table} meets the required format, or it stops with an error message if any checks fail.
}
\description{
This function checks if an annotation table meets the required format for metabolite annotation. It verifies the presence of necessary columns, data types, and ensures that no critical values are missing (NA).
}
\examples{
\dontrun{
# Assuming `my_annotation_table` is the table to validate:
check_annotation_table(my_annotation_table)
}


}
