% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-utils.R
\name{check_column_name}
\alias{check_column_name}
\title{Check Whether a Column Name Is Available}
\usage{
check_column_name(df, column.name)
}
\arguments{
\item{df}{A data frame.}

\item{column.name}{A proposed column name.}
}
\value{
A character scalar containing an available column name.
}
\description{
Returns \code{column.name} if it is not already present in \code{df}.
If the name is already used, a numeric suffix is appended to create a
unique column name.
}
\examples{
df <-
  data.frame(x = 1:2, y = 2:3, y.1 = 3:4)
df

check_column_name(df = df, column.name = "y")
check_column_name(df = df, column.name = "y.1")
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
