% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checking.R
\name{check_mass_dataset}
\alias{check_mass_dataset}
\title{Check Mass Dataset}
\usage{
check_mass_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{A data frame containing expression data.}

\item{sample_info}{A data frame containing sample information.}

\item{variable_info}{A data frame containing variable information.}

\item{sample_info_note}{A data frame containing notes for sample information.}

\item{variable_info_note}{A data frame containing notes for variable information.}
}
\value{
A string indicating the result of the checks. It returns "all good." if all checks pass, otherwise it returns an error message.
}
\description{
This function performs various checks on the mass dataset to ensure its integrity and consistency. It checks the format and content of \code{expression_data}, \code{sample_info}, \code{variable_info}, \code{sample_info_note}, and \code{variable_info_note}.
}
\details{
The function performs a series of checks to validate the integrity and consistency of the mass dataset. It checks for missing data, data types, and the relationship between different components of the dataset.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
  check_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
