% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checking.R
\name{check_tidymass_class_data}
\alias{check_tidymass_class_data}
\title{check_tidymass_class_data}
\usage{
check_tidymass_class_data(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{expression_data.}

\item{sample_info}{sample_info.}

\item{variable_info}{variable_info.}

\item{sample_info_note}{sample_info_note.}

\item{variable_info_note}{variable_info_note.}
}
\value{
Notice of data checking.
}
\description{
Check data format.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
