% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_class.R
\docType{methods}
\name{colnames,tidymass-method}
\alias{colnames,tidymass-method}
\alias{colnames}
\title{colnames}
\usage{
\S4method{colnames}{tidymass}(x)
}
\arguments{
\item{x}{object}
}
\description{
colnames
}
