% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_msdial2mass_dataset.R
\name{convert_msdial2mass_dataset}
\alias{convert_msdial2mass_dataset}
\title{Convert MS-Dial Data to mass_dataset Object}
\usage{
convert_msdial2mass_dataset(x, path = ".", file_name)
}
\arguments{
\item{x}{A data frame containing MS-Dial data. If missing, the function will read from a file.}

\item{path}{The directory where the file is located. Default is the current directory.}

\item{file_name}{The name of the file to read if \code{x} is missing.}
}
\value{
A \code{mass_dataset} object containing the processed MS-Dial data.
}
\description{
This function converts data from MS-Dial into a \code{mass_dataset} object.
It reads the data either from a given data frame or from a file and processes it to create a \code{mass_dataset} object.
}
\details{
The function reads MS-Dial data and processes it to create a \code{mass_dataset} object.
It extracts sample information, variable information, and expression data.
It also performs checks to ensure the data is correctly formatted.
}
\examples{
data("msdial_table")
object <- convert_msdial2mass_dataset(x = msdial_table)
dim(extract_expression_data(object))
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
