% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_class.R
\name{create_tidymass_class}
\alias{create_tidymass_class}
\title{create_tidymass_class}
\usage{
create_tidymass_class(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{MS1 peak table name. Columns are samples and rows are variables.}

\item{sample_info}{Sample information name.}

\item{variable_info}{MS1 peak table name. Columns are samples and rows are variables.}

\item{sample_info_note}{Sample information name.}

\item{variable_info_note}{Sample information name.}
}
\value{
A tidymass-class object.
}
\description{
Create tidymass object.
}
\examples{
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_tidymass_class(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
  object
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
