% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{dim.mass_dataset}
\alias{dim.mass_dataset}
\title{dim}
\usage{
\method{dim}{mass_dataset}(x)
}
\arguments{
\item{x}{A mass_dataset object.}
}
\value{
A named integer vector with the dimensions of the mass_dataset object.
The names of the dimensions are "variables" and "samples".
}
\description{
dim
}
