% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{dimnames.mass_dataset}
\alias{dimnames.mass_dataset}
\title{dimnames}
\usage{
\method{dimnames}{mass_dataset}(x)
}
\arguments{
\item{x}{A \code{mass_dataset} object whose \code{expression_data} slot's dimension
names you want to retrieve.}
}
\value{
The dimension names (row and column names) of
the \code{expression_data} slot in the \code{mass_dataset} object.
}
\description{
dimnames
}
