% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mass_dataset.R
\name{export_mass_dataset}
\alias{export_mass_dataset}
\title{Export mass_dataset Object to Files}
\usage{
export_mass_dataset(
  object,
  file_type = c("csv", "xlsx"),
  ms2_file_type = c("msp", "mgf"),
  path = "."
)
}
\arguments{
\item{object}{A \code{mass_dataset} object to be exported.}

\item{file_type}{The type of file to export the data to. Options are "csv" or "xlsx". Default is "csv".}

\item{ms2_file_type}{The type of file to export the MS2 data to. Options are "msp" or "mgf". Default is "msp".}

\item{path}{The directory where the exported files will be saved. Default is the current directory.}
}
\value{
No return value. The function writes files to disk.
}
\description{
Exports the contents of a \code{mass_dataset} object to files. The files can be in CSV or XLSX format.
Additionally, MS2 data can be exported in MSP or MGF format.
}
\examples{
\dontrun{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
export_mass_dataset(object, file_type = "csv", path = tempdir())
}

}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
