% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mass_dataset.R
\name{export_ms2_data}
\alias{export_ms2_data}
\title{Export MS2 Data from mass_dataset Object}
\usage{
export_ms2_data(object, file_type = c("mgf", "msp"), path = ".")
}
\arguments{
\item{object}{A \code{mass_dataset} object containing MS2 data to be exported.}

\item{file_type}{The type of file to export the MS2 data to. Options are "mgf" or "msp". Default is "mgf".}

\item{path}{The directory where the exported MS2 data files will be saved. Default is the current directory.}
}
\value{
No return value. The function writes MS2 data files to disk.
Returns NULL and a warning if no MS2 data is present in the object.
}
\description{
Exports the MS2 data contained in a \code{mass_dataset} object to files.
The files can be in MGF or MSP format.
}
\examples{
\dontrun{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
object <- mutate_ms2(
  object = object,
  column = "rp",
  polarity = "positive",
  path = system.file("ms2_data", package = "massdataset")
)
export_ms2_data(object, file_type = "mgf", path = tempdir())
}


}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
