% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tidymass_class.R
\name{export_tidymass_class}
\alias{export_tidymass_class}
\title{Export tidymass class object to csv/xlsx files}
\usage{
export_tidymass_class(object, file_type = c("csv", "xlsx"), path = ".")
}
\arguments{
\item{object}{(required) tidymass class object.}

\item{file_type}{(required) csv or xlsx}

\item{path}{work directory.}
}
\description{
Export tidymass class object to csv/xlsx files
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
