% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_in_package.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{Expression data}
\format{
A data frame with 1000 rows (1000 variables) and 8 columns (8 samples).
}
\description{
expression_data
A data frame contains the expression values for each variable in each sample.
}
\examples{
data(expression_data)
}
\keyword{data}
