% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\docType{methods}
\name{extract_expression_data}
\alias{extract_expression_data}
\alias{extract_sample_info}
\alias{extract_variable_info}
\alias{extract_annotation_table}
\alias{extract_variable_info_note}
\alias{extract_sample_info_note}
\alias{extract_process_info}
\alias{extract_ms2_data}
\alias{sample_info,mass_dataset-method}
\alias{expression_data,mass_dataset-method}
\alias{variable_info,mass_dataset-method}
\alias{process_info,mass_dataset-method}
\alias{ms2_data,mass_dataset-method}
\alias{sample_info_note,mass_dataset-method}
\alias{variable_info_note,mass_dataset-method}
\alias{annotation_table,mass_dataset-method}
\title{Extract expression_data from mass_dataset}
\usage{
extract_expression_data(object)

extract_sample_info(object)

extract_variable_info(object)

extract_annotation_table(object)

extract_variable_info_note(object)

extract_sample_info_note(object)

extract_process_info(object)

extract_ms2_data(object)

\S4method{sample_info}{mass_dataset}(object)

\S4method{expression_data}{mass_dataset}(object)

\S4method{variable_info}{mass_dataset}(object)

\S4method{process_info}{mass_dataset}(object)

\S4method{ms2_data}{mass_dataset}(object)

\S4method{sample_info_note}{mass_dataset}(object)

\S4method{variable_info_note}{mass_dataset}(object)

\S4method{annotation_table}{mass_dataset}(object)
}
\arguments{
\item{object}{(required) mass_dataset class object}
}
\value{
A data frame (expression_data).

A data frame (sample_info).

A data frame (variable_info).

A data frame (annotation_table).

A data frame (variable_info_note).

A data frame (sample_info_note).

A list of tidymass_parameter.

A ms2_data class object.

A data.frame (sample_info)

A data.frame (expression_data)

A data.frame (variable_info)

A list of tidymass_parameter class objects

A ms2_data class object.

A data.frame (sample_info_note)

A data.frame (variable_info_note)

A data.frame (annotation_table)
}
\description{
Extract expression_data from mass_dataset

Extract sample information.

Extract variable information.

Extract variable information.

Extract variable information note.

Extract sample information note.

Extract process information

Extract ms2 data
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    variable_info = variable_info,
    sample_info = sample_info
  )
 expression_data2 =
 extract_expression_data(object = object)
 head(expression_data2)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 sample_info2 =
 extract_sample_info(object = object)
 head(sample_info2)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 variable_info2 =
 extract_variable_info(object = object)
 head(variable_info2)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 annotation_table =
 extract_annotation_table(object = object)
 head(annotation_table)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 test =
 extract_variable_info_note(object = object)
 head(test)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 test =
 extract_sample_info_note(object = object)
 head(test)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object
extract_process_info(object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 test =
 extract_ms2_data(object = object)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
