% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_class.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,tidymass-method}
\title{Extract parts of tidymass}
\usage{
x[i,j,drop = FALSE,...]
}
\arguments{
\item{x}{tidymass class object}

\item{i}{row index}

\item{j}{column index}

\item{...}{other parameters}

\item{drop}{drop or not.}
}
\description{
Extract parts of tidymass
}
