% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_annotation_table}
\alias{extract_annotation_table}
\title{Extract Annotation Table from mass_dataset Object}
\usage{
extract_annotation_table(object, with_expression_data = FALSE)
}
\arguments{
\item{object}{An object of class "mass_dataset" from which the annotation table will be extracted.}

\item{with_expression_data}{Logical, whether to include expression data in the output. Default is FALSE.}
}
\value{
A data frame containing the annotation table, optionally including expression data.
}
\description{
This function extracts the annotation table from an object of class "mass_dataset". It first checks if the object belongs to the class "mass_dataset" and then extracts the annotation table. Optionally, it can also include expression data.
}
\details{
The function uses the \code{check_object_class} function to ensure that the object belongs to the class "mass_dataset". It then extracts the '@annotation_table' slot and converts it to a data frame. If \code{with_expression_data} is TRUE, it will also join the expression data.
}
\examples{
\dontrun{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
annotation_table =
  extract_annotation_table(object = object)
head(annotation_table)
}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
