% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_expression_data}
\alias{extract_expression_data}
\title{extract_expression_data}
\usage{
extract_expression_data(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A data frame.
}
\description{
Extract expression data.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_tidymass_class(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 expression_data2 =
 extract_expression_data(object = object)
 head(expression_data2)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
