% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_ms2_data}
\alias{extract_ms2_data}
\title{Extract ms2 data}
\usage{
extract_ms2_data(object)
}
\arguments{
\item{object}{(required) mass_dataset class object.}
}
\value{
A ms2_data class object.
}
\description{
Extract ms2 data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
