% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_sample_info}
\alias{extract_sample_info}
\title{Extract Sample Information from mass_dataset Object}
\usage{
extract_sample_info(object)
}
\arguments{
\item{object}{An object of class "mass_dataset" from which the sample information will be extracted.}
}
\value{
A data frame containing the sample information.
}
\description{
This function extracts the sample information from an object of class "mass_dataset". It first checks if the object belongs to the class "mass_dataset" and then extracts the sample information.
}
\details{
The function uses the \code{check_object_class} function to ensure that the object belongs to the class "mass_dataset". It then extracts the '@sample_info' slot and converts it to a data frame.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 sample_info2 =
 extract_sample_info(object = object)
 head(sample_info2)
}
\seealso{
\code{\link{check_object_class}}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
