% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_sample_info_note}
\alias{extract_sample_info_note}
\title{extract_sample_info_note}
\usage{
extract_sample_info_note(object)
}
\arguments{
\item{object}{mass_dataset class object.}
}
\value{
A data frame.
}
\description{
Extract sample information note.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
