% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_variable_info}
\alias{extract_variable_info}
\title{Extract Variable Information from mass_dataset Object}
\usage{
extract_variable_info(object, with_expression_data = FALSE)
}
\arguments{
\item{object}{An object of class "mass_dataset" from which the variable information will be extracted.}

\item{with_expression_data}{Logical, whether to include expression data in the output. Default is FALSE.}
}
\value{
A data frame containing the variable information, optionally including expression data.
}
\description{
This function extracts the variable information from an object of class "mass_dataset". It first checks if the object belongs to the class "mass_dataset" and then extracts the variable information. Optionally, it can also include expression data.
}
\details{
The function uses the \code{check_object_class} function to ensure that the object belongs to the class "mass_dataset". It then extracts the '@variable_info' slot and converts it to a data frame. If the object has an '@annotation_table', it will join this information with the variable information. If \code{with_expression_data} is TRUE, it will also join the expression data.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 variable_info2 =
 extract_variable_info(object = object)
 head(variable_info2)
}
\seealso{
\code{\link{check_object_class}}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
