% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_filtering.R
\name{filter_samples}
\alias{filter_samples}
\title{filter_samples}
\usage{
filter_samples(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A tidymass-class object
}
\description{
Number of samples
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
