% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_filtering.R
\name{filter_variables}
\alias{filter_variables}
\title{Filter variables based on the conditions}
\usage{
filter_variables(
  object,
  flist,
  prune = FALSE,
  apply_to = "all",
  according_to_samples = "all"
)
}
\arguments{
\item{object}{(required) tidymass class object.}

\item{flist}{(required) A function or list of functions that take a vector
of abundance values and return a logical.}

\item{prune}{(optional) A logical. Default \code{FALSE}. If \code{TRUE}, then
the function returns the pruned tidymass-class object, rather
than the logical vector of variables that passed the filter.}

\item{apply_to}{what variables you want to apply this function.
Default is "all". If you
only want to apply to specific variables,
please set it as a vector of sample names. Other
variables will be set as TRUE.}

\item{according_to_samples}{What samples used to filter variables.
Default is "all". If you
want to use only several samples, provide they names as a vector.}
}
\value{
A logical vector equal to the number of variables in tidymass-class.
Alternatively, if \code{prune==TRUE}, the pruned tidymass-class
object is returned instead.
}
\description{
Filter variables based on the conditions
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_tidymass_class(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

####Filter variables which have more than 50\% MVs in all samples.
library(tidyverse)
filter_variables(object, function(x) {sum(is.na(x))/length(x) < 0.5}) \%>\% 
head()
filter_variables(object, function(x) {sum(is.na(x))/length(x) < 0.5}, 
prune = TRUE)

####Filter variables which have more than 50\% MVs in only QC samples.
filter_variables(object, flist = function(x) {sum(is.na(x))/length(x) < 0.5}, 
                 prune = TRUE, 
             according_to_samples = 
           get_sample_id(object)[extract_sample_info(object)$class == "QC"])

####Filter variables which have more than 50\% MVs in QC or subject samples.
idx1 = 
filter_variables(object, flist = function(x) {sum(is.na(x))/length(x) < 0.5}, 
                 according_to_samples = 
         get_sample_id(object)[extract_sample_info(object)$class == "QC"])

idx2 = 
  filter_variables(object, 
  flist = function(x) {sum(is.na(x))/length(x) < 0.5}, 
       according_to_samples = 
   get_sample_id(object)[extract_sample_info(object)$class == "Subject"])
idx =
  which(idx1 | idx2)

object2 = object[idx,]
object2
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
