% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-massdataset_logo.R
\name{get_massdataset_version}
\alias{get_massdataset_version}
\title{Retrieve the Version of MassDataset Package}
\usage{
get_massdataset_version()
}
\value{
A character string indicating the version of the massdataset package.
}
\description{
This function retrieves the current version of the massdataset package installed in the R environment.
}
\examples{
get_massdataset_version()
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
