% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{get_mv_number}
\alias{get_mv_number}
\title{get_mv_number}
\usage{
get_mv_number(
  object,
  by = c("total", "sample", "variable"),
  show_by = c("number", "percentage")
)
}
\arguments{
\item{object}{mass_dataset class object.}

\item{by}{total: Missing value number in total. sample: Missing value number in each sample.
variable: Missing value number in each variable.}

\item{show_by}{number: missing value number. percentage: ratio.}
}
\value{
A numeric vector
}
\description{
Get missing value number/percentage in expression
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
object 
head(get_variable_id(object = object))
get_mv_number(object)
get_mv_number(object, by = "sample")
head(get_mv_number(object, by = "variable", "percentage"))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
