% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{get_variable_id}
\alias{get_variable_id}
\title{get_variable_id}
\usage{
get_variable_id(object)
}
\arguments{
\item{object}{mass_dataset class object.}
}
\value{
A character vector
}
\description{
Get sample names
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
library(tidyverse)
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_variable_id(object = object) \%>\% head()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
