% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mass_dataset.R
\docType{methods}
\name{ggplot_mass_dataset}
\alias{ggplot_mass_dataset}
\title{Create a ggplot object from mass_dataset class}
\usage{
ggplot_mass_dataset(
  object,
  direction = c("variable", "sample"),
  variable_index = 1,
  variable_id,
  sample_index = 1,
  sample_id
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{direction}{variable or sample}

\item{variable_index}{variable index}

\item{variable_id}{variable ID}

\item{sample_index}{sample index}

\item{sample_id}{sample ID}
}
\value{
A ggplot class
}
\description{
Create a ggplot object from mass_dataset class
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    variable_info = variable_info,
    sample_info = sample_info
  )
library(ggplot2)
ggplot_mass_dataset(object, direction = "variable", 
                    variable_index = 1) +
  geom_boxplot(aes(x = class, color = class)) +
  geom_jitter(aes(x = class, color = class))

object \%>\% 
  `+`(1) \%>\% 
  log(10) \%>\% 
  scale() \%>\% 
ggplot_mass_dataset(direction = "sample", 
                    sample_index = 2) +
  geom_boxplot(aes(x = 1)) +
  geom_jitter(aes(x = 1, color = mz))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
