% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{head.mass_dataset}
\alias{head.mass_dataset}
\title{head}
\usage{
\method{head}{mass_dataset}(x, ...)
}
\arguments{
\item{x}{A \code{mass_dataset} object.}

\item{...}{Additional parameters.}
}
\value{
A \code{mass_dataset} object containing the first rows of \code{expression_data}
and the corresponding linked metadata.
}
\description{
head
}
