% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Print “hello world”}
\usage{
hello_world()
}
\value{
Invisibly returns the string "hello world".
}
\description{
A minimal demo function that prints “hello world” to the console.
}
\examples{
hello_world()
}
