% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_plot.R
\name{intensity_plot}
\alias{intensity_plot}
\title{Create Intensity Plot for a Variable in a mass_dataset Object}
\usage{
intensity_plot(
  object,
  variable_id,
  variable_index,
  color_by,
  order_by,
  desc = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{object}{A mass_dataset object.}

\item{variable_id}{A character string specifying the ID of the variable of interest.
If provided, 'variable_index' will be ignored.}

\item{variable_index}{An integer specifying the index of the variable of interest.
If not provided, you can use 'variable_id' to specify the variable by its ID.}

\item{color_by}{A character string specifying the attribute to color data points.
Default is "no", meaning no coloring.}

\item{order_by}{A character string specifying the attribute to order the samples.
Default is "sample_id". If set to "na", no specific order is applied.}

\item{desc}{A logical value indicating whether to order the samples in descending order.
Default is FALSE.}

\item{interactive}{A logical value indicating whether to create an interactive plot using plotly.
Default is TRUE.}
}
\value{
A ggplot2 or plotly plot object.
}
\description{
This function generates an intensity plot for a specific variable in a mass_dataset object.
You can specify the variable either by its ID or its index. Additionally, you can customize
the color of data points based on a specific attribute and order the samples accordingly.
}
\examples{
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object

intensity_plot(object = object, variable_index = 1)
intensity_plot(object = object, variable_index = 1, color_by = "class")
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
