% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{intersect,mass_dataset,mass_dataset-method}
\alias{intersect,mass_dataset,mass_dataset-method}
\title{intersect}
\usage{
\S4method{intersect}{mass_dataset,mass_dataset}(x, y)
}
\arguments{
\item{x}{A mass_dataset object.}

\item{y}{Another mass_dataset object.}
}
\value{
A vector containing the common sample IDs between the
two mass_dataset objects.
}
\description{
intersect
}
