% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{is.na.mass_dataset}
\alias{is.na.mass_dataset}
\title{is.na}
\usage{
\method{is.na}{mass_dataset}(x)
}
\arguments{
\item{x}{A \code{mass_dataset} object whose \code{expression_data}
slot you want to check for missing values.}
}
\value{
A logical matrix of the same dimensions as \code{expression_data},
indicating the presence of missing values.
}
\description{
is.na
}
