% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_left_join.R
\name{left_join_mass_dataset}
\alias{left_join_mass_dataset}
\title{Left join for mass_dataset}
\usage{
left_join_mass_dataset(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)
}
\arguments{
\item{x}{(required) mass_dataset class object.}

\item{y}{A data frame.}

\item{by}{by, see ?_left_join}

\item{copy}{copy, see ?_left_join}

\item{suffix}{suffix, see ?_left_join}

\item{...}{Other arguments, see ?_left_join}

\item{keep}{keep, see ?_left_join}
}
\value{
a mass_dataset class object.
}
\description{
Left join for mass_dataset
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
