% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{length.mass_dataset}
\alias{length.mass_dataset}
\title{length}
\usage{
\method{length}{mass_dataset}(x)
}
\arguments{
\item{x}{A \code{mass_dataset} object whose \code{expression_data} slot's length
you want to find.}
}
\value{
The length of the \code{expression_data} slot in the
\code{mass_dataset} object.
}
\description{
length
}
