% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-mass_dataset_class.R
\docType{class}
\name{mass_dataset-class}
\alias{mass_dataset-class}
\title{S4 Class Representing a mass_dataset}
\description{
The \code{mass_dataset} class is designed to hold various types of data
related to mass spectrometry experiments, including expression data,
MS2 data, annotation tables, and other related information.
}
\section{Slots}{

\describe{
\item{\code{expression_data}}{A data.frame containing the expression data.}

\item{\code{ms2_data}}{A list containing MS2 data.}

\item{\code{annotation_table}}{A data.frame containing annotation information.}

\item{\code{sample_info}}{A data.frame containing sample information.}

\item{\code{variable_info}}{A data.frame containing variable information.}

\item{\code{sample_info_note}}{A data.frame containing notes about the sample information.}

\item{\code{variable_info_note}}{A data.frame containing notes about the variable information.}

\item{\code{process_info}}{A list containing information about the data processing steps.}

\item{\code{other_files}}{A list containing other related files.}

\item{\code{version}}{A character string indicating the version of the dataset.}

\item{\code{activated}}{A character string indicating which slot is currently activated for manipulation.}
}}

\examples{
\dontrun{
data("expression_data")
data("sample_info")
data("variable_info")
new_data <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
new_data
}

}
