% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{get_sample_number}
\alias{get_sample_number}
\alias{get_variable_number}
\alias{get_sample_id}
\alias{get_variable_id}
\alias{get_mv_number}
\title{Get Some Information of a mass_dataset Object}
\usage{
get_sample_number(object)

get_variable_number(object)

get_sample_id(object)

get_variable_id(object)

get_mv_number(
  object,
  by = c("total", "sample", "variable"),
  show_by = c("number", "percentage")
)
}
\arguments{
\item{object}{A mass_dataset object.}

\item{by}{A character string specifying how to count missing values.
Options are "total", "sample", or "variable". Default is "total".}

\item{show_by}{A character string specifying the format to show the result.
Options are "number" or "percentage". Default is "number".}
}
\value{
the number of columns (samples) in expression_data

the number of rows (variables) in expression_data

sample IDs from expression_data column names

variable IDs from expression_data row names
}
\description{
These functions returns the information of a mass_dataset object.
}
\examples{
###get_sample_number
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
 get_sample_number(object = object)
###get_variable_number
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
 get_variable_number(object = object)
###get_sample_id
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
 get_sample_id(object = object)
###get_variable_id
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
 head(get_variable_id(object = object))
###get_mv_number
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
object
head(get_variable_id(object = object))
get_mv_number(object)
get_mv_number(object, by = "sample")
head(get_mv_number(object, by = "variable", "percentage"))
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
