% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{class}
\name{mass_dataset}
\alias{mass_dataset}
\title{An S4 class that stores the MS dataset}
\description{
An S4 class that stores the MS dataset
}
\section{Slots}{

\describe{
\item{\code{expression_data}}{expression_data}

\item{\code{ms2_data}}{ms2_data}

\item{\code{sample_info}}{sample_info}

\item{\code{variable_info}}{variable_info}

\item{\code{sample_info_note}}{sample_info_note}

\item{\code{variable_info_note}}{variable_info_note}

\item{\code{process_info}}{process_info}

\item{\code{version}}{version}

\item{\code{activated}}{activated}
}}

