% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-utils.R
\name{massdataset_packages}
\alias{massdataset_packages}
\title{List all packages in the massdataset}
\usage{
massdataset_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical; if \code{TRUE}, include \code{massdataset} itself in the
returned package names.}
}
\value{
A character vector of package names used by the \code{massdataset}
workflow.
}
\description{
List all packages in the massdataset
}
\examples{
massdataset_packages()
}
