% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masstools_compat.R, R/masstools_match_mz_rt.R
\name{match_mz_rt.data.frame}
\alias{match_mz_rt.data.frame}
\alias{match_mz_rt.matrix}
\alias{match_mz_rt}
\title{Match Features by m/z and Retention Time}
\usage{
\method{match_mz_rt}{data.frame}(
  data1,
  data2,
  mz.tol,
  rt.tol = 30,
  rt.error.type = c("relative", "abs")
)

\method{match_mz_rt}{matrix}(
  data1,
  data2,
  mz.tol,
  rt.tol = 30,
  rt.error.type = c("relative", "abs")
)

match_mz_rt(
  data1,
  data2,
  mz.tol = 10,
  rt.tol = 30,
  rt.error.type = c("abs", "relative")
)
}
\arguments{
\item{data1}{A \code{mass_dataset} object.}

\item{data2}{A second \code{mass_dataset} object or compatible structure.}

\item{mz.tol}{Numeric m/z tolerance for matching.}

\item{rt.tol}{Numeric retention time tolerance for matching.}

\item{rt.error.type}{Retention time error type.}
}
\value{
A data frame of candidate matches, or \code{NULL} if no matches are found.
}
\description{
Match Features by m/z and Retention Time

Match Features by m/z and Retention Time
}
