% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mass_dataset.R
\docType{methods}
\name{merge_mass_dataset}
\alias{merge_mass_dataset}
\title{merge_mass_dataset}
\usage{
merge_mass_dataset(
  x,
  y,
  sample_direction = c("left", "right", "full", "inner"),
  variable_direction = c("left", "right", "full", "inner"),
  sample_by = c("sample_id"),
  variable_by = c("variable_id", "mz", "rt")
)
}
\arguments{
\item{x}{(required) A mass_dataset class object.}

\item{y}{(required) A mass_dataset class object.}

\item{sample_direction}{How to merge samples, should be
left, right, inner or full. See ?left_join}

\item{variable_direction}{How to merge variables,
should be left, right, inner or full.}

\item{sample_by}{merge samples by what columns from sample_info.}

\item{variable_by}{merge variables by what columns from variable_info}
}
\value{
A merged mass_dataset.
}
\description{
Merge two mass_dataset objects. More information can be found
here \url{https://www.tidymass.org/massdataset/articles/process_info.html}
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

x <- object[1:3, 5:7]
y <- object[2:4, 6:8]

z1 <- merge_mass_dataset(
  x = x,
  y = y,
  sample_direction = "full",
  variable_direction = "full"
)

z2 <- merge_mass_dataset(
  x = x,
  y = y,
  sample_direction = "inner",
  variable_direction = "full"
)

dim(extract_expression_data(z1))
dim(extract_expression_data(z2))

z3 <- merge_mass_dataset(
  x = object[1:3, 5:7],
  y = object[4:6, 6:8],
  sample_direction = "full",
  variable_direction = "full"
)
dim(extract_expression_data(z3))
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
