% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_data-methods.R
\name{ms2_plot,mass_dataset-method}
\alias{ms2_plot,mass_dataset-method}
\title{ms2_plot}
\usage{
\S4method{ms2_plot}{mass_dataset}(
  spectrum1,
  spectrum2,
  spectrum1_name = "spectrum1",
  spectrum2_name = "spectrum2",
  range.mz,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity",
  col1 = "red",
  col2 = "black",
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  legend.title.size = 15,
  legend.text.size = 15,
  interactive_plot = FALSE
)
}
\arguments{
\item{spectrum1}{ms2_data}

\item{spectrum2}{Spectrum 2.}

\item{spectrum1_name}{Name of spectrum1}

\item{spectrum2_name}{Name of spectrum2}

\item{range.mz}{range.mz}

\item{ppm.tol}{ppm.tol}

\item{mz.ppm.thr}{mz.ppm.thr}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{col1}{Color 1.}

\item{col2}{Color 2.}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size.}

\item{legend.title.size}{legend.title.size}

\item{legend.text.size}{legend.text.size}

\item{interactive_plot}{Interactive plot or not.}
}
\value{
ms2 matching plot
}
\description{
ms2_plot
}
