% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R
\name{ms2_data}
\alias{ms2_data}
\title{Generic Function for Retrieving MS2 Spectra Data}
\usage{
ms2_data(object, ...)
}
\arguments{
\item{object}{An object from which MS2 spectra data will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a matrix, dataframe, or list with the MS2 spectra information.
}
\description{
The \code{ms2_data} generic function retrieves mass spectrometry (MS2) spectra data from the provided object.
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
