% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_ms2_plot.R
\name{ms2_plot}
\alias{ms2_plot}
\title{Plot MS2 Spectra}
\usage{
ms2_plot(
  spectrum1,
  spectrum2,
  spectrum1_name = "spectrum1",
  spectrum2_name = "spectrum2",
  range.mz,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  xlab = "Mass to charge ratio (m/z)",
  ylab = "Relative intensity",
  col1 = "red",
  col2 = "black",
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  legend.title.size = 15,
  legend.text.size = 15,
  interactive_plot = FALSE
)
}
\arguments{
\item{spectrum1}{An \code{ms2_data} object.}

\item{spectrum2}{An optional second \code{ms2_data} object for comparison.}

\item{spectrum1_name}{Label for the first spectrum.}

\item{spectrum2_name}{Label for the second spectrum.}

\item{range.mz}{m/z plotting range.}

\item{ppm.tol}{PPM tolerance for matching peaks.}

\item{mz.ppm.thr}{m/z threshold used with ppm matching.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{col1}{Color for the first spectrum.}

\item{col2}{Color for the second spectrum.}

\item{title.size}{Title text size.}

\item{lab.size}{Axis label text size.}

\item{axis.text.size}{Axis tick text size.}

\item{legend.title.size}{Legend title text size.}

\item{legend.text.size}{Legend text size.}

\item{interactive_plot}{Whether to create an interactive plot.}
}
\description{
Plot MS2 Spectra
}
