% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{msdial_table}
\alias{msdial_table}
\title{msdial_table feature table}
\format{
A data frame.
}
\usage{
msdial_table
}
\description{
msdial_table
An example feature table from mzMine.
}
\keyword{datasets}
