% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_annotation_table.R
\name{mutate_annotation_table}
\alias{mutate_annotation_table}
\title{Update Annotation Table in a mass_dataset Object}
\usage{
mutate_annotation_table(object, annotation_table)
}
\arguments{
\item{object}{A \code{mass_dataset} object containing the existing annotation table.}

\item{annotation_table}{A data frame containing new annotation data to be added to the \code{mass_dataset} object. The table should contain specific columns as detailed below.}
}
\value{
The function returns the updated \code{mass_dataset} object with the appended and processed annotation table.
}
\description{
This function updates the annotation table in a \code{mass_dataset} object by appending new annotation data and retaining only the top annotations based on the total score.
}
\details{
This function allows you to update or append new annotations to the existing annotation table in a \code{mass_dataset} object. The function ensures that the new annotations are properly arranged based on the \code{Total.score}, and it retains only the top annotations for each \code{variable_id} based on the specified \code{candidate.num}.

The \code{annotation_table} must contain the following columns:
\describe{
\item{variable_id}{The unique identifier for each variable.}
\item{ms2_files_id}{The ID of the MS2 file associated with each annotation.}
\item{ms2_spectrum_id}{The ID of the MS2 spectrum associated with each annotation.}
\item{Compound.name}{The name of the annotated compound.}
\item{CAS.ID}{The CAS ID of the compound.}
\item{HMDB.ID}{The HMDB ID of the compound.}
\item{KEGG.ID}{The KEGG ID of the compound.}
\item{Lab.ID}{The lab identifier for the compound.}
\item{Adduct}{The adduct form of the compound.}
\item{mz.error}{The error in m/z matching.}
\item{mz.match.score}{The score for the m/z match.}
\item{RT.error}{The error in retention time matching.}
\item{RT.match.score}{The score for the retention time match.}
\item{CE}{Collision energy used in MS2 matching.}
\item{SS}{Spectral similarity score for MS2 matching.}
\item{Total.score}{The total score for the annotation.}
\item{Database}{The database used for annotation.}
\item{Level}{The confidence level of the annotation.}
}
}
\examples{
\dontrun{
# Example usage
updated_object <- mutate_annotation_table(
  object = my_dataset,
  annotation_table = new_annotation_data
)
}


}
