% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_rsd.R
\name{mutate_rsd}
\alias{mutate_rsd}
\title{Mutate Relative Standard Deviation (RSD) in mass_dataset Object}
\usage{
mutate_rsd(object, according_to_samples = "all")
}
\arguments{
\item{object}{A mass_dataset object.}

\item{according_to_samples}{A character vector specifying the sample IDs to consider when
calculating the RSD. Default is "all", which considers all samples.}
}
\value{
A modified mass_dataset object with an updated \code{variable_info} slot.
}
\description{
This function adds a new column to the \code{variable_info} slot of a mass_dataset object,
which contains the relative standard deviation (RSD) for each variable according to the samples specified.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

object2 <- mutate_rsd(object = object)
head(extract_variable_info(object2))

qc_samples <- get_sample_id(object)[extract_sample_info(object)$class == "QC"]
object3 <- mutate_rsd(object = object2, according_to_samples = qc_samples)
head(extract_variable_info(object3))
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
