% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sample_na.R
\name{mutate_sample_na_number}
\alias{mutate_sample_na_number}
\title{Mutate Sample NA Count in mass_dataset Object}
\usage{
mutate_sample_na_number(object, according_to_variables = "all")
}
\arguments{
\item{object}{A mass_dataset object.}

\item{according_to_variables}{A character vector specifying the variable IDs to consider when
calculating the count of NA values. Default is "all", which considers all variables.}
}
\value{
A modified mass_dataset object with an updated \code{sample_info} slot.
}
\description{
This function adds a new column to the \code{sample_info} slot of a mass_dataset object,
which contains the count of NA (Not Available) values for each sample according to the variables specified.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object

##calculate NA number according to all the variables
object2 =
  mutate_sample_na_number(object = object)

colnames(extract_sample_info(object))
colnames(extract_sample_info(object2))
extract_sample_info_note(object2)

##calculate NA number according to only variables with mz > 100
variable_id =
object2 \%>\%
  activate_mass_dataset(what = "variable_info") \%>\%
  filter(mz > 100) \%>\%
  pull(variable_id)

object3 =
  mutate_sample_na_number(object = object2,
                according_to_variables = variable_id)

object3

head(extract_sample_info(object3))
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
