% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{mzmine_table}
\alias{mzmine_table}
\title{mzmine feature table}
\format{
A data frame.
}
\usage{
mzmine_table
}
\description{
mzmine_table
An example feature table from mzMine.
}
\keyword{datasets}
