% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{names.mass_dataset}
\alias{names.mass_dataset}
\title{names}
\usage{
\method{names}{mass_dataset}(x)
}
\arguments{
\item{x}{A \code{mass_dataset} object whose \code{expression_data}
slot's names you want to retrieve.}
}
\value{
The names of the elements in the \code{expression_data}
slot of the \code{mass_dataset} object.
}
\description{
names
}
