% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{ncol.mass_dataset}
\alias{ncol.mass_dataset}
\title{ncol}
\usage{
\method{ncol}{mass_dataset}(x)
}
\arguments{
\item{x}{A mass_dataset object.}
}
\value{
A named integer representing the number of samples in the mass_dataset object.
The name of the integer is "samples".
}
\description{
ncol
}
