% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{ncol,mass_dataset-method}
\alias{ncol,mass_dataset-method}
\alias{ncol}
\title{ncol}
\usage{
\S4method{ncol}{mass_dataset}(x)
}
\arguments{
\item{x}{object}
}
\description{
ncol
}
