% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{nrow.mass_dataset}
\alias{nrow.mass_dataset}
\title{nrow}
\usage{
\method{nrow}{mass_dataset}(x)
}
\arguments{
\item{x}{A mass_dataset object.}
}
\value{
A named integer representing the number of variables in the mass_dataset object.
The name of the integer is "variables".
}
\description{
nrow
}
