% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{nrow,mass_dataset-method}
\alias{nrow,mass_dataset-method}
\alias{nrow}
\title{nrow}
\usage{
\S4method{nrow}{mass_dataset}(x)
}
\arguments{
\item{x}{object}
}
\description{
nrow
}
