% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R
\name{process_info}
\alias{process_info}
\title{Generic Function for Retrieving Process Information}
\usage{
process_info(object, ...)
}
\arguments{
\item{object}{An object from which process information will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a dataframe or a list describing the processing details.
}
\description{
The \code{process_info} generic function retrieves detailed information related to the data processing of the provided object.
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
