% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{cbind.mass_dataset}
\alias{cbind.mass_dataset}
\alias{rbind.mass_dataset}
\alias{head.mass_dataset}
\alias{tail.mass_dataset}
\title{cbind}
\usage{
\method{cbind}{mass_dataset}(..., deparse.level = 1)

\method{rbind}{mass_dataset}(..., deparse.level = 1)

\method{head}{mass_dataset}(x, ...)

\method{tail}{mass_dataset}(x, ...)
}
\arguments{
\item{...}{mass_datasets}

\item{deparse.level}{deparse.level}

\item{x}{x}
}
\value{
mass_dataset

mass_dataset

mass_dataset class object

mass_dataset class object
}
\description{
cbind

rbind

head

tail
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
