% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{rbind_mass_dataset}
\alias{rbind_mass_dataset}
\title{Row-wise Binding of mass_dataset Objects}
\usage{
rbind_mass_dataset(x, y, deparse.level = 1)
}
\arguments{
\item{x}{A mass_dataset object. If missing, the function returns \code{y}.}

\item{y}{A mass_dataset object. If missing, the function returns \code{x}.}

\item{deparse.level}{Not used, for compatibility only.}
}
\value{
A new mass_dataset object that is the row-wise combination of \code{x} and \code{y}.
}
\description{
This function combines two mass_dataset objects by rows. It checks for
compatibility in terms of sample information, variable information, and other
attributes before performing the row-wise binding.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
x <- object[1:3, 1:3]
y <- object[4:6, 1:3]

result <- rbind_mass_dataset(x, y)
dim(result)

}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
