% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_arrange.R, R/dplyr_count.R,
%   R/dplyr_filter.R, R/dplyr_glimpse.R, R/dplyr_group_by.R,
%   R/dplyr_left_join.R, R/dplyr_mutate.R, R/dplyr_pull.R, R/dplyr_relocate.R,
%   R/dplyr_rename.R, R/dplyr_select.R, R/dplyr_slice.R, R/dplyr_summarize.R,
%   R/ggplot2_ggplot.R, R/ms2_data-methods.R, R/reexports.R, R/tidyr_drop_na.R,
%   R/tidyr_fill.R, R/tidyr_pivot_longer.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{desc}
\alias{count}
\alias{filter}
\alias{glimpse}
\alias{group_by}
\alias{left_join}
\alias{mutate}
\alias{transmute}
\alias{mutate_all}
\alias{mutate_at}
\alias{n}
\alias{pull}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{select}
\alias{slice}
\alias{slice_head}
\alias{slice_tail}
\alias{slice_sample}
\alias{slice_min}
\alias{slice_max}
\alias{summarize}
\alias{ggplot}
\alias{aes}
\alias{\%>\%}
\alias{case_when}
\alias{summarise}
\alias{drop_na}
\alias{fill}
\alias{pivot_longer}
\title{Objects exported from other packages}
\value{
\code{NULL} No return value. This page is only for documentation purposes.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{glimpse}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr]{transmute}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggplot}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tidyr}{\code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{fill}}, \code{\link[tidyr]{pivot_longer}}}
}}

