% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{rownames.mass_dataset}
\alias{rownames.mass_dataset}
\title{rownames}
\usage{
\method{rownames}{mass_dataset}(x, do.NULL = TRUE, prefix = "row")
}
\arguments{
\item{x}{A mass_dataset object.}

\item{do.NULL}{Passed through for compatibility with \code{base::rownames()}.}

\item{prefix}{Passed through for compatibility with \code{base::rownames()}.}
}
\value{
A character vector containing the row names (variable IDs) of
the expression_data slot in the mass_dataset object.
}
\description{
rownames
}
