% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_class.R
\docType{methods}
\name{rownames,tidymass-method}
\alias{rownames,tidymass-method}
\alias{rownames}
\title{rownames}
\usage{
\S4method{rownames}{tidymass}(x)
}
\arguments{
\item{x}{object}
}
\description{
rownames
}
