% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R, R/5-data_in_package.R
\docType{data}
\name{sample_info_note}
\alias{sample_info_note}
\title{Generic Function for Retrieving Sample Information Notes}
\format{
A data frame with 4 rows (4 columns in sample_info) and 2 columns.
}
\usage{
sample_info_note(object, ...)
}
\arguments{
\item{object}{An object from which sample information notes will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a character vector, list, or dataframe containing notes or metadata about the sample information.
}
\description{
The \code{sample_info_note} generic function retrieves notes or metadata associated with sample information from the provided object.

sample_info_note
A data frame contains description for each sample information column.
}
\examples{
data(sample_info_note)
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
\keyword{data}
