% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_data_class.R, R/tidymass_parameter_class.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
\S4method{show}{ms2_data}(object)

\S4method{show}{tidymass_parameter}(object)
}
\arguments{
\item{object}{A \code{tidymass_parameter} instance.}
}
\value{
message

message
}
\description{
show method for ms2_data

show method for tidymass_parameter
}
\author{
Xiaotao Shen

Xiaotao Shen
}
