% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_mz_rt_plot.R
\docType{methods}
\name{show_mz_rt_plot}
\alias{show_mz_rt_plot}
\title{Generate mz-rt Plot for mass_dataset Object}
\usage{
show_mz_rt_plot(object, hex = FALSE)
}
\arguments{
\item{object}{A \code{mass_dataset} object for which the mz-rt plot is to be generated.}

\item{hex}{Logical. If TRUE, a hexbin plot is generated; otherwise, a scatter plot is generated. Default is FALSE.}
}
\value{
A ggplot object representing the mz-rt plot.
}
\description{
This function generates a plot of mass-to-charge ratio (m/z) against retention time (rt) for a \code{mass_dataset} object. The plot can be either a hexbin plot or a scatter plot based on the \code{hex} parameter.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

show_mz_rt_plot(object)
show_mz_rt_plot(log(object, 10))
show_mz_rt_plot(object, hex = TRUE)
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
