% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_explore_plot.R
\docType{methods}
\name{show_variable_missing_values}
\alias{show_variable_missing_values}
\title{show_variable_missing_values}
\usage{
show_variable_missing_values(
  object,
  color_by,
  order_by,
  percentage = FALSE,
  show_x_text = FALSE,
  show_x_ticks = FALSE,
  desc = FALSE
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{color_by}{which column (in variable_info) is used to color variables}

\item{order_by}{which column (in variable_info) is used to order variables}

\item{percentage}{show MV percentage? TRUE or FALSE}

\item{show_x_text}{show x axis text or not? TRUE or FALSE}

\item{show_x_ticks}{show x ticks or not? TRUE or FALSE}

\item{desc}{descend sample order or not. TRUE or FALSE.}
}
\value{
A ggplot2 object
}
\description{
show missing values for each variable
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

show_variable_missing_values(object)
show_variable_missing_values(object, color_by = "mz")
show_variable_missing_values(object, color_by = "rt") +
  ggplot2::scale_color_gradient(low = "skyblue", high = "red")

show_variable_missing_values(object, color_by = "mz",
                             order_by = "na")
show_variable_missing_values(object, color_by = "mz",
                             order_by = "na",
                           desc = TRUE, percentage = TRUE)
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
