% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-mass_dataset.R
\name{[.mass_dataset}
\alias{[.mass_dataset}
\alias{$.mass_dataset}
\title{[ method}
\usage{
\method{[}{mass_dataset}(x, i, j, drop = FALSE, ...)

\method{$}{mass_dataset}(x, name)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{...}{Other parameters}

\item{name}{A \link{name} or a string.}
}
\value{
mass_dataset
}
\description{
[ method

$ method
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
