% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summarize_samples.R
\name{summarise_samples}
\alias{summarise_samples}
\alias{summarize_samples}
\alias{summarise_samples.mass_dataset}
\title{merge samples}
\usage{
summarise_samples(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  group_by,
  ...
)

summarize_samples(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  group_by,
  ...
)

\method{summarise_samples}{mass_dataset}(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  group_by,
  ...
)
}
\arguments{
\item{object}{A \code{mass_dataset} object.}

\item{what}{Summary statistic used to combine samples within each group.}

\item{group_by}{Character vector giving one or more columns in \code{sample_info}
used to define the grouping structure.}

\item{...}{Additional arguments passed to downstream methods.}
}
\value{
A \code{mass_dataset} object with grouped samples collapsed into summary
profiles and updated \code{sample_info}.
}
\description{
merge samples
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)

object2 <- summarise_samples(
  object,
  what = "mean_intensity",
  group_by = "class"
)
head(extract_expression_data(object2))
head(extract_sample_info(object2))
}
