% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{summary.mass_dataset}
\alias{summary.mass_dataset}
\title{summary}
\usage{
\method{summary}{mass_dataset}(object, ...)
}
\arguments{
\item{object}{A mass_dataset object.}

\item{...}{Additional arguments to be passed to the \code{summary}
function for the expression data.}
}
\value{
summary of the expression_data slot
}
\description{
summary
}
