% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{tail.mass_dataset}
\alias{tail.mass_dataset}
\title{tail}
\usage{
\method{tail}{mass_dataset}(x, ...)
}
\arguments{
\item{x}{A \code{mass_dataset} object.}

\item{...}{Additional parameters.}
}
\value{
A \code{mass_dataset} object containing the last rows of \code{expression_data}
and the corresponding linked metadata.
}
\description{
tail
}
