% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_parameter-methods.R
\name{translate_tidymass_parameter}
\alias{translate_tidymass_parameter}
\title{Translate tidymass_parameter Object to Data Frame}
\usage{
translate_tidymass_parameter(object)
}
\arguments{
\item{object}{A \code{tidymass_parameter} object that you want to translate.}
}
\value{
A data frame containing the package name, function name, parameters, and time of the \code{tidymass_parameter} object.
}
\description{
This function takes a \code{tidymass_parameter} object and translates its slots into a data frame.
It is particularly useful for converting the parameter information into a more readable format.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
object2 <- mutate_mean_intensity(object)
translate_tidymass_parameter(extract_process_info(object2)[[1]])
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
