% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mass_dataset.R
\name{update_mass_dataset}
\alias{update_mass_dataset}
\title{Update a mass_dataset Object}
\usage{
update_mass_dataset(object)
}
\arguments{
\item{object}{A \code{mass_dataset} object that you want to update.}
}
\value{
A \code{mass_dataset} object with updated slots and process information.
}
\description{
This function updates the slots of a \code{mass_dataset} object to ensure consistency
among \code{sample_info}, \code{variable_info}, and \code{expression_data}. It also logs the
update process.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object

qc_object <- object[, extract_sample_info(object)$class == "QC"]
qc_object <- update_mass_dataset(object = qc_object)
dim(qc_object)

feature_object <-
  object[extract_variable_info(object)$mz < 300, ]
feature_object <- update_mass_dataset(object = feature_object)
dim(feature_object)
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
