% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tidymass_class.R
\name{update_tidymass_class}
\alias{update_tidymass_class}
\title{Update tidymass class object after give the one new item}
\usage{
update_tidymass_class(object)
}
\arguments{
\item{object}{(required) tidymass class object.}
}
\value{
A logical vector equal to the number of variables in tidymass-class.
Alternatively, if \code{prune==TRUE}, the pruned tidymass-class
object is returned instead.
}
\description{
Update tidymass class object after give the one new item
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_tidymass_class(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

####only remain QC samples
library(tidyverse)
sample_info = extract_sample_info(object = object)

sample_info =
  sample_info \%>\%
  dplyr::filter(class == "QC")

object@sample_info = sample_info

object = update_tidymass_class(object = object)
object

####only remain feature with mz < 300
variable_info = extract_variable_info(object = object)

variable_info =
  variable_info \%>\%
  dplyr::filter(mz < 300)

object@variable_info = variable_info

object = update_tidymass_class(object = object)
object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
