% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mass_dataset.R
\name{update_variable_info}
\alias{update_variable_info}
\title{Update Variable Information in a mass_dataset Object}
\usage{
update_variable_info(object)
}
\arguments{
\item{object}{A \code{mass_dataset} object whose \code{variable_info} and \code{variable_info_note} slots you want to update.}
}
\value{
A \code{mass_dataset} object with updated \code{variable_info} and \code{variable_info_note} slots.
}
\description{
This function updates the \code{variable_info} and \code{variable_info_note} slots of a \code{mass_dataset} object.
It ensures that the columns in \code{variable_info} are consistent with the names in \code{variable_info_note}.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object <- create_mass_dataset(
  expression_data = expression_data,
  sample_info = sample_info,
  variable_info = variable_info
)
updated_md <- update_variable_info(object)
head(extract_variable_info(updated_md))
head(extract_variable_info_note(updated_md))

}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
