% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R, R/5-data_in_package.R
\docType{data}
\name{variable_info}
\alias{variable_info}
\title{Generic Function for Retrieving Variable Information}
\format{
A data frame with 1000 rows (1000 variables) and 3 columns.
}
\usage{
variable_info(object, ...)
}
\arguments{
\item{object}{An object from which variable information will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a dataframe or a list describing the variables.
}
\description{
The \code{variable_info} generic function retrieves detailed information related to variables in the provided object.

variable_info
A data frame contains the variable information.
}
\examples{
data(variable_info)
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
\keyword{data}
