% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsSet-class.R
\docType{class}
\name{xcmsSet-class}
\alias{xcmsSet-class}
\title{An S4 class that stores the MS dataset from xcms (Prof. Steffen Neumann)}
\description{
An S4 class that stores the MS dataset from xcms (Prof. Steffen Neumann)
}
\section{Slots}{

\describe{
\item{\code{peaks}}{peaks}

\item{\code{groups}}{groups}

\item{\code{groupidx}}{groupidx}

\item{\code{filled}}{filled}

\item{\code{phenoData}}{phenoData}

\item{\code{rt}}{rt}

\item{\code{filepaths}}{filepaths}

\item{\code{profinfo}}{profinfo}

\item{\code{dataCorrection}}{dataCorrection}

\item{\code{polarity}}{polarity}

\item{\code{progressInfo}}{progressInfo}

\item{\code{progressCallback}}{progressCallback}

\item{\code{mslevel}}{mslevel}

\item{\code{scanrange}}{scanrange}

\item{\code{.processHistory}}{.processHistory}
}}

