% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-activate_mass_dataset.R
\name{activate_mass_dataset}
\alias{activate_mass_dataset}
\title{Activate a Specific Dataset}
\usage{
activate_mass_dataset(
  .data,
  what = c("sample_info", "variable_info", "expression_data", "annotation_table")
)
}
\arguments{
\item{.data}{A \code{mass_dataset} object.}

\item{what}{A character string specifying which dataset to activate.
The available options are:
\itemize{
\item "sample_info": Activates the sample information dataset.
\item "variable_info": Activates the variable information dataset.
\item "expression_data": Activates the expression data.
\item "annotation_table": Activates the annotation table.
}
By default, it activates the "sample_info" dataset.}
}
\value{
Returns the \code{mass_dataset} object with the specified dataset activated.
}
\description{
Activates a dataset within the mass dataset object for further analysis.
}
\details{
This function allows users to activate a specific dataset within the
\code{mass_dataset} object. The activated dataset can be then used in subsequent
analysis steps.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
object
object@activated
object =
  activate_mass_dataset(.data = object, what = "sample_info")
object@activated
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
