% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_mass_dataset.R
\name{activate_mass_dataset}
\alias{activate_mass_dataset}
\title{Determine the context of subsequent manipulations}
\usage{
activate_mass_dataset(
  .data,
  what = c("sample_info", "variable_info", "expression_data", "sample_info_note",
    "variable_info_note")
)
}
\arguments{
\item{.data}{A mass_dataset class object}

\item{what}{What should get activated? Possible values are \code{sample_info},
\code{expression_data}, \code{variable_info}, \code{sample_info_note} or
\code{variable_info_note}.}
}
\value{
A tidyclas class object
}
\description{
Determine the context of subsequent manipulations
}
\examples{
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object
object@activated
object =
  activate_mass_dataset(.data = object, what = "sample_info")
object@activated
}
