% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic-mass_dataset.R
\name{log.mass_dataset}
\alias{log.mass_dataset}
\alias{abs.mass_dataset}
\alias{sqrt.mass_dataset}
\alias{scale.mass_dataset}
\alias{+,mass_dataset,numeric-method}
\alias{-,mass_dataset,numeric-method}
\alias{*,mass_dataset,numeric-method}
\alias{/,mass_dataset,numeric-method}
\alias{colSums.mass_dataset}
\alias{rowSums.mass_dataset}
\alias{colMeans.mass_dataset}
\alias{rowMeans.mass_dataset}
\title{log}
\usage{
\method{log}{mass_dataset}(x, base = exp(1))

\method{abs}{mass_dataset}(x)

\method{sqrt}{mass_dataset}(x)

\method{scale}{mass_dataset}(x, center = TRUE, scale = TRUE)

\S4method{+}{mass_dataset,numeric}(e1, e2)

\S4method{-}{mass_dataset,numeric}(e1, e2)

\S4method{*}{mass_dataset,numeric}(e1, e2)

\S4method{/}{mass_dataset,numeric}(e1, e2)

\method{colSums}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\method{rowSums}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\method{colMeans}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\method{rowMeans}{mass_dataset}(x, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{x}

\item{base}{base}

\item{center}{center}

\item{scale}{scale}

\item{e1}{a mass_dataset class object}

\item{e2}{numeric}

\item{na.rm}{na.rm}

\item{dims}{dims}
}
\value{
mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

vector object

vector object

vector object

vector object
}
\description{
log

abs

sqrt

scale

\itemize{
\item 
}

\itemize{
\item 
}

\itemize{
\item 
}

/

colSums

rowSums

colMeans

rowMeans
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
