% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-mass_dataset.R
\name{$.mass_dataset}
\alias{$.mass_dataset}
\title{$ method}
\usage{
\method{$}{mass_dataset}(x, name)
}
\arguments{
\item{x}{A mass_dataset class}

\item{name}{A \link{name} or a string.}
}
\description{
$ method of mass_dataset
}
