% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{cbind.mass_dataset}
\alias{cbind.mass_dataset}
\title{cbind}
\usage{
\method{cbind}{mass_dataset}(..., deparse.level = 1)
}
\arguments{
\item{...}{One or more \code{mass_dataset} objects to be column bound.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. Default is 1.}
}
\value{
mass_dataset
}
\description{
cbind
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
