% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{cbind_mass_dataset}
\alias{cbind_mass_dataset}
\title{cbind mass_data class}
\usage{
cbind_mass_dataset(x, y, deparse.level = 1)
}
\arguments{
\item{x}{A \code{mass_dataset} object.}

\item{y}{Another \code{mass_dataset} object to be column bound with \code{x}.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. Default is 1.}
}
\value{
mass_dataset
}
\description{
cbind mass_data class
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
