% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{cbind_mass_dataset}
\alias{cbind_mass_dataset}
\title{cbind mass_data class}
\usage{
cbind_mass_dataset(x, y, deparse.level = 1)
}
\arguments{
\item{x}{A \code{mass_dataset} object.}

\item{y}{Another \code{mass_dataset} object to be column bound with \code{x}.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. Default is 1.}
}
\value{
mass_dataset
}
\description{
cbind mass_data class
}
\examples{
library(magrittr)

# Create first mass_dataset object
expression_data_x <- data.frame(
  sample1 = c(1.1, 2.2, 3.3),
  sample2 = c(1.5, 2.5, 3.5),
  row.names = c("feature1", "feature2", "feature3")
)

sample_info_x <- data.frame(
  sample_id = c("sample1", "sample2"),
  batch = c(1, 1),
  age = c(30, 32),
  class = c("QC", "QC"),
  row.names = c("sample1", "sample2")
)

sample_info_note_x <- data.frame(
  name = colnames(sample_info_x),
  meaning = c("Sample ID", "Batch ID", "Age", "Sample Class"),
  row.names = colnames(sample_info_x)
)

variable_info <- data.frame(
  variable_id = c("feature1", "feature2", "feature3"),
  row.names = c("feature1", "feature2", "feature3")
)

variable_info_note <- data.frame(
  name = colnames(variable_info),
  meaning = "Feature ID",
  row.names = colnames(variable_info)
)

mass_x <- new(
  Class = "mass_dataset",
  expression_data = expression_data_x,
  sample_info = sample_info_x,
  sample_info_note = sample_info_note_x,
  variable_info = variable_info,
  variable_info_note = variable_info_note,
  ms2_data = list(),
  annotation_table = data.frame(),
  process_info = list(),
  other_files = list(),
  version = "1.0.0",
  activated = "expression_data"
)

# Create second mass_dataset object
expression_data_y <- data.frame(
  sample1 = c(1.3, 2.3, 3.4),
  sample2 = c(1.7, 2.7, 3.8),
  row.names = c("feature1", "feature2", "feature3")
)

sample_info_y <- data.frame(
  sample_id = c("sample1", "sample2"),
  batch = c(2, 2),
  age = c(31, 34),
  class = c("Subject", "Subject"),
  row.names = c("sample1", "sample2")
)

sample_info_note_y <- data.frame(
  name = colnames(sample_info_y),
  meaning = c("Sample ID", "Batch ID", "Age", "Sample Class"),
  row.names = colnames(sample_info_y)
)

mass_y <- new(
  Class = "mass_dataset",
  expression_data = expression_data_y,
  sample_info = sample_info_y,
  sample_info_note = sample_info_note_y,
  variable_info = variable_info,
  variable_info_note = variable_info_note,
  ms2_data = list(),
  annotation_table = data.frame(),
  process_info = list(),
  other_files = list(),
  version = "1.0.0",
  activated = "expression_data"
)

# Combine the two objects
combined_object <- cbind_mass_dataset(mass_x, mass_y)

# View result
combined_object@expression_data

}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
