% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_column_name}
\alias{check_column_name}
\title{check_column_name}
\usage{
check_column_name(df, column.name)
}
\arguments{
\item{df}{A data frame.}

\item{column.name}{new column name}
}
\value{
A column name
}
\description{
check_column_name
}
\examples{
df <- 
  data.frame(x = 1:2, y = 2:3, y.1 = 3:4)
df

check_column_name(df = df, column.name = "y")
check_column_name(df = df, column.name = "y.1")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
