% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checking.R
\docType{methods}
\name{check_mass_dataset}
\alias{check_mass_dataset}
\title{check_mass_dataset}
\usage{
check_mass_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{expression_data.}

\item{sample_info}{sample_info.}

\item{variable_info}{variable_info.}

\item{sample_info_note}{sample_info_note.}

\item{variable_info_note}{variable_info_note.}
}
\value{
Notice of data checking.
}
\description{
Check data format fpr mass_dataset class object.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
  check_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
