% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{check_mass_dataset_class}
\alias{check_mass_dataset_class}
\title{check_mass_dataset_class}
\usage{
check_mass_dataset_class(object)
}
\arguments{
\item{object}{mass_dataset class.}
}
\value{
Notice of data checking.
}
\description{
Check mass_dataset class object.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object <-
  create_mass_dataset(expression_data = expression_data,
                      sample_info = sample_info,
                      variable_info = variable_info)
  check_mass_dataset_class(object)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
