% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checking.R
\name{check_object_class}
\alias{check_object_class}
\title{check_object_class}
\usage{
check_object_class(
  object,
  class = c("mass_dataset", "tidymass_parameter", "databaseClass", "metIdentifyClass")
)
}
\arguments{
\item{object}{object}

\item{class}{object should be what class}
}
\value{
Notice of data checking.
}
\description{
check_object_class
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
check_object_class(object, "mass_dataset")
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
