% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checking.R
\name{check_object_class}
\alias{check_object_class}
\title{Check Object Class}
\usage{
check_object_class(
  object,
  class = c("mass_dataset", "tidymass_parameter", "databaseClass", "metIdentifyClass",
    "ms2_data")
)
}
\arguments{
\item{object}{The object to be checked.}

\item{class}{A character vector specifying the class or classes to check against. Default options are "mass_dataset", "tidymass_parameter", "databaseClass", "metIdentifyClass", and "ms2_data".}
}
\value{
None. The function stops execution and throws an error if the object does not belong to the specified class.
}
\description{
This function checks if the provided object belongs to a specified class. It supports multiple classes including "mass_dataset", "tidymass_parameter", "databaseClass", "metIdentifyClass", and "ms2_data".
}
\details{
The function uses the \code{is} function to check if the object belongs to the specified class. If the object does not belong to the class, the function stops and returns an error message.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
check_object_class(object, "mass_dataset")
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
