% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\name{colnames.mass_dataset}
\alias{colnames.mass_dataset}
\title{colnames}
\usage{
\method{colnames}{mass_dataset}(x)
}
\arguments{
\item{x}{A mass_dataset object.}
}
\value{
A character vector containing the column names (sample IDs) of the
expression_data slot in the mass_dataset object.
}
\description{
colnames
}
