% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{colnames,mass_dataset-method}
\alias{colnames,mass_dataset-method}
\alias{colnames}
\title{colnames}
\usage{
\S4method{colnames}{mass_dataset}(x)
}
\arguments{
\item{x}{object}
}
\description{
colnames
}
