% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mass_dataset2mztab.R
\name{convert_mass_dataset2mztab}
\alias{convert_mass_dataset2mztab}
\title{convert_mass_dataset2mztab}
\usage{
convert_mass_dataset2mztab(object, path = ".")
}
\arguments{
\item{object}{mass_dataset class object.}

\item{path}{The path to save the mzTab-m file.}
}
\value{
A mzTab-m file.
}
\description{
Convert mass_dataset class to mzTab-m format.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
