% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mass_dataset2summarizedexperiment.R
\name{convert_mass_dataset2summarizedexperiment}
\alias{convert_mass_dataset2summarizedexperiment}
\title{Convert mass_dataset to SummarizedExperiment Object}
\usage{
convert_mass_dataset2summarizedexperiment(object)
}
\arguments{
\item{object}{A \code{mass_dataset} object containing \code{sample_info}, \code{variable_info}, and \code{expression_data}.}
}
\value{
A \code{SummarizedExperiment} object containing the same data as the input \code{mass_dataset} object.
}
\description{
This function converts a \code{mass_dataset} object to a \code{SummarizedExperiment} object.
It extracts the \code{sample_info}, \code{variable_info}, and \code{expression_data} from the \code{mass_dataset} object
and constructs a new \code{SummarizedExperiment} object.
}
\details{
The function checks if the input object is of class \code{mass_dataset}. If it is, it extracts the \code{sample_info},
\code{variable_info}, and \code{expression_data} and uses these to create a new \code{SummarizedExperiment} object.
}
\examples{
library(magrittr)

# Create expression data
expression_data_x <- data.frame(
  sample1 = c(1.1, 2.2, 3.3),
  sample2 = c(1.5, 2.5, 3.5),
  row.names = c("feature1", "feature2", "feature3")
)

# Sample info
sample_info_x <- data.frame(
  sample_id = c("sample1", "sample2"),
  batch = c(1, 1),
  age = c(30, 32),
  class = c("QC", "QC"),
  row.names = c("sample1", "sample2")
)

# Sample info note
sample_info_note_x <- data.frame(
  name = colnames(sample_info_x),
  meaning = c("Sample ID", "Batch ID", "Age", "Sample Class"),
  row.names = colnames(sample_info_x)
)

# Variable info
variable_info_x <- data.frame(
  variable_id = c("feature1", "feature2", "feature3"),
  row.names = c("feature1", "feature2", "feature3")
)

# Variable info note
variable_info_note_x <- data.frame(
  name = colnames(variable_info_x),
  meaning = "Feature ID",
  row.names = colnames(variable_info_x)
)

# Construct mass_dataset object
mass_dataset_object <- new(
  Class = "mass_dataset",
  expression_data = expression_data_x,
  sample_info = sample_info_x,
  sample_info_note = sample_info_note_x,
  variable_info = variable_info_x,
  variable_info_note = variable_info_note_x,
  ms2_data = list(),
  annotation_table = data.frame(),
  process_info = list(),
  other_files = list(),
  version = "1.0.0",
  activated = "expression_data"
)

# Convert to SummarizedExperiment
convert_mass_dataset2summarizedexperiment(mass_dataset_object)

}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
