% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mass_dataset2summarizedexperiment.R
\name{convert_mass_dataset2summarizedexperiment}
\alias{convert_mass_dataset2summarizedexperiment}
\title{convert_mass_dataset2summarizedexperiment}
\usage{
convert_mass_dataset2summarizedexperiment(object)
}
\arguments{
\item{object}{mass_dataset class object.}
}
\value{
A SummarizedExperiment file.
}
\description{
Convert mass_dataset class to SummarizedExperiment from
SummarizedExperiment package.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
