% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_msdial2mass_dataset.R
\docType{methods}
\name{convert_msdial2mass_dataset}
\alias{convert_msdial2mass_dataset}
\title{convert_msdial2mass_dataset}
\usage{
convert_msdial2mass_dataset(x)
}
\arguments{
\item{x}{Feature table from MS-DIAL.}

\item{rt_unit}{RT unit (minute or second) in feature table.}
}
\value{
A mass_dataset-class object.
}
\description{
Convert MS-DIAL feature table to mass_dataset class.
}
\examples{

data("msdial_table")
object =
  convert_msdial2mass_dataset(x = msdial_table)
object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
