% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mztab2mass_dataset.R
\name{convert_mztab2mass_dataset}
\alias{convert_mztab2mass_dataset}
\title{convert_mztab2mass_dataset}
\usage{
convert_mztab2mass_dataset(file, path = ".")
}
\arguments{
\item{file}{mzTab-m file}

\item{path}{where is the mzTab-m file.}
}
\value{
A mass_dataset-class object.
}
\description{
Convert mzTab-m format to mass_dataset class.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
