% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mzmine2mass_dataset.R
\docType{methods}
\name{convet_mzmine2mass_dataset}
\alias{convet_mzmine2mass_dataset}
\title{convet_mzmine2mass_dataset}
\usage{
convet_mzmine2mass_dataset(x, rt_unit = c("minute", "second"))
}
\arguments{
\item{x}{Feature table from mzmine.}

\item{rt_unit}{RT unit (minute or second) in feature table.}
}
\value{
A mass_dataset-class object.
}
\description{
Convert mzmine feature table to mass_dataset class.
}
\examples{
data("mzmine_table")
head(mzmine_table[,1:3])
object =
  convet_mzmine2mass_dataset(x = mzmine_table)
object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
