% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\name{create_mass_dataset}
\alias{create_mass_dataset}
\title{create_mass_dataset}
\usage{
create_mass_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{MS1 peak table name. Columns are samples and rows are variables.}

\item{sample_info}{Sample information name.}

\item{variable_info}{MS1 peak table name. Columns are samples and rows are variables.}

\item{sample_info_note}{Sample information name.}

\item{variable_info_note}{Sample information name.}
}
\value{
A mass_dataset-class object.
}
\description{
Create mass_dataset object.
}
\examples{
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
