% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{create_mass_dataset}
\alias{create_mass_dataset}
\title{Create a New mass_dataset Object}
\usage{
create_mass_dataset(
  expression_data,
  sample_info,
  variable_info,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{MS1 peak table name.
\url{https://tidymass.github.io/massdataset/articles/data_import_and_export.html}}

\item{sample_info}{Sample information name.
\url{https://tidymass.github.io/massdataset/articles/data_import_and_export.html}}

\item{variable_info}{MS1 peak table name.
Columns are samples and rows are variables.
\url{https://tidymass.github.io/massdataset/articles/data_import_and_export.html}}

\item{sample_info_note}{Sample information name.
\url{https://tidymass.github.io/massdataset/articles/data_import_and_export.html}}

\item{variable_info_note}{Sample information name.
\url{https://tidymass.github.io/massdataset/articles/data_import_and_export.html}}
}
\value{
An object of class \code{mass_dataset}.
}
\description{
This function creates a new \code{\link[massdataset]{mass_dataset}} object
by taking in various data frames and performing checks on them.
}
\examples{
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
