% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mass_dataset.R
\name{export_mass_dataset}
\alias{export_mass_dataset}
\title{Export mass_dataset class object to csv/xlsx files}
\usage{
export_mass_dataset(object, file_type = c("csv", "xlsx"), path = ".")
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{file_type}{(required) csv or xlsx}

\item{path}{work directory.}
}
\description{
Export mass_dataset class object to csv/xlsx files
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
