% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_metdna.R
\name{export_mass_dataset4metdna}
\alias{export_mass_dataset4metdna}
\title{export_mass_dataset4metdna}
\usage{
export_mass_dataset4metdna(object, path = ".")
}
\arguments{
\item{object}{mass_dataset class object.}

\item{path}{The path to save the mzTab-m file.}
}
\value{
csv files
}
\description{
Export mass_dataset to MetDNA require data format.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
