% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mass_dataset.R
\docType{methods}
\name{export_ms2_data}
\alias{export_ms2_data}
\title{export_ms2_data}
\usage{
export_ms2_data(object, file_type = c("mgf", "msp"), path = ".")
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{file_type}{(required) mgf, msp}

\item{path}{(required) work directory.}
}
\value{
mgf, msp files
}
\description{
Export mass_dataset's ms2_data to mgf/msp
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
