% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mass_dataset.R
\name{export_ms2_data}
\alias{export_ms2_data}
\title{Export MS2 Data from mass_dataset Object}
\usage{
export_ms2_data(object, file_type = c("mgf", "msp"), path = ".")
}
\arguments{
\item{object}{A \code{mass_dataset} object containing MS2 data to be exported.}

\item{file_type}{The type of file to export the MS2 data to. Options are "mgf" or "msp". Default is "mgf".}

\item{path}{The directory where the exported MS2 data files will be saved. Default is the current directory.}
}
\value{
No return value. The function writes MS2 data files to disk.
Returns NULL and a warning if no MS2 data is present in the object.
}
\description{
Exports the MS2 data contained in a \code{mass_dataset} object to files.
The files can be in MGF or MSP format.
}
\examples{
\dontrun{
# Assuming 'data' is a mass_dataset object with MS2 data
export_ms2_data(data, file_type = "mgf", path = "output/")
}


}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
