% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R, R/5-data_in_package.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{Generic Function for Retrieving Expression Data}
\format{
A data frame with 1000 rows (1000 variables) and 8 columns (8 samples).
}
\usage{
expression_data(object, ...)
}
\arguments{
\item{object}{An object from which expression data will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a matrix or dataframe of expression values.
}
\description{
The \code{expression_data} generic function retrieves gene expression-related data from the provided object.

expression_data
A data frame contains the expression values for each variable in each sample.
}
\examples{
data(expression_data)
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
\keyword{data}
