% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,mass_dataset-method}
\title{Extract parts of mass_dataset}
\usage{
x[i,j,drop = FALSE,...]
}
\arguments{
\item{x}{mass_dataset class object}

\item{i}{row index}

\item{j}{column index}

\item{...}{other parameters}

\item{drop}{drop or not.}
}
\description{
Extract parts of mass_dataset
}
