% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_annotation_table}
\alias{extract_annotation_table}
\title{extract_annotation_table}
\usage{
extract_annotation_table(object)
}
\arguments{
\item{object}{mass_dataset class object.}
}
\value{
A data frame.
}
\description{
Extract variable information.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 annotation_table =
 extract_annotation_table(object = object)
 head(annotation_table)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
