% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_expression_data}
\alias{extract_expression_data}
\title{extract_expression_data}
\usage{
extract_expression_data(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A data frame.
}
\description{
Extract expression data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
