% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_expression_data}
\alias{extract_expression_data}
\title{Extract Expression Data from mass_dataset Object}
\usage{
extract_expression_data(object)
}
\arguments{
\item{object}{An object of class "mass_dataset" from which the expression data will be extracted.}
}
\value{
A data frame containing the expression data.
}
\description{
This function extracts the expression data from an object of class "mass_dataset". It first checks if the object belongs to the class "mass_dataset" and then extracts the expression data.
}
\details{
The function uses the \code{check_object_class} function to ensure that the object belongs to the class "mass_dataset". It then extracts the '@expression_data' slot and converts it to a data frame.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    variable_info = variable_info,
    sample_info = sample_info
  )
 expression_data2 =
 extract_expression_data(object = object)
 head(expression_data2)
}
\seealso{
\code{\link{extract_sample_info}}
\code{\link{extract_variable_info}}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
