% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_ms2_data}
\alias{extract_ms2_data}
\title{Extract MS2 Data from mass_dataset Object}
\usage{
extract_ms2_data(object)
}
\arguments{
\item{object}{An object of class "mass_dataset" from which the MS2 data will be extracted.}
}
\value{
A list containing the MS2 data.
}
\description{
This function extracts the MS2 data from an object of class "mass_dataset". It first checks if the object belongs to the class "mass_dataset" and then extracts the MS2 data.
}
\details{
The function uses the \code{check_object_class} function to ensure that the object belongs to the class "mass_dataset". It then extracts the '@ms2_data' slot and returns it as a list.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 test =
 extract_ms2_data(object = object)
}
\seealso{
\code{\link{check_object_class}}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
