% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_process_info}
\alias{extract_process_info}
\title{Extract process information}
\usage{
extract_process_info(object)
}
\arguments{
\item{object}{(required) tidymass class object.}
}
\value{
A data.frame.
}
\description{
Extract process information
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
library(massdataset)

object =
  create_tidymass_class(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object
extract_process_info(object)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
