% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_process_info}
\alias{extract_process_info}
\title{Extract Process Information from mass_dataset Object}
\usage{
extract_process_info(object)
}
\arguments{
\item{object}{An object of class "mass_dataset" from which the process information will be extracted.}
}
\value{
A list containing the process information.
}
\description{
This function extracts the process information from an object of class "mass_dataset". It first checks if the object belongs to the class "mass_dataset" and then extracts the process information.
}
\details{
The function uses the \code{check_object_class} function to ensure that the object belongs to the class "mass_dataset". It then extracts the '@process_info' slot and returns it as a list.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object
extract_process_info(object)
}
\seealso{
\code{\link{check_object_class}}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
