% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_sample_info}
\alias{extract_sample_info}
\title{extract_sample_info}
\usage{
extract_sample_info(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A data frame.
}
\description{
Extract sample information.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
