% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R
\name{extract_variable_info}
\alias{extract_variable_info}
\title{extract_variable_info}
\usage{
extract_variable_info(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A data frame.
}
\description{
Extract variable information.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
