% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_filtering.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples based on the conditions}
\usage{
filter_samples(object, flist, prune = TRUE, apply_to = "all")
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{flist}{(required) A function or list of functions that take a vector
of abundance values and return a logical.}

\item{prune}{(optional) A logical. Default \code{FALSE}. If \code{TRUE}, then
the function returns the pruned mass_dataset-class object, rather
than the logical vector of samples that passed the filter.}

\item{apply_to}{what samples you want to apply this function. default is "all". If you
only want to apply to specific samples, please set it as a vector of sample names. Other
samples will be set as TRUE.}
}
\value{
A logical vector equal to the number of samples in mass_dataset-class.
Alternatively, if \code{prune==TRUE}, the pruned mass_dataset-class
object is returned instead.
}
\description{
Filter samples based on the conditions
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 
filter_samples(object, function(x) {
  sum(is.na(x)) / length(x) < 0.4
})

filter_samples(object, function(x) {
  sum(is.na(x)) / length(x) < 0.4
}, prune = FALSE)

##only apply to Subject sample
object2 =
filter_samples(
 object = object,
 flist = function(x) {
   sum(is.na(x))/length(x) < 0.2
 },
 prune = TRUE,
 apply_to = get_sample_id(object)[extract_sample_info(object)$class == "Subject"]
)

object2
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
