% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdataset_logo.R
\docType{methods}
\name{get_massdataset_version}
\alias{get_massdataset_version}
\title{get_massdataset_version}
\usage{
get_massdataset_version()
}
\value{
version
}
\description{
Get massdataset package version
}
\examples{
get_massdataset_version()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
