% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{get_sample_id}
\alias{get_sample_id}
\title{get_sample_id}
\usage{
get_sample_id(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A character vector
}
\description{
Get sample names
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_tidymass_class(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_sample_id(object = object)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
