% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{get_sample_number}
\alias{get_sample_number}
\title{get_sample_number}
\usage{
get_sample_number(object)
}
\arguments{
\item{object}{mass_dataset class object.}
}
\value{
A numeric.
}
\description{
Number of samples
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_sample_number(object = object)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
