% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{get_variable_id}
\alias{get_variable_id}
\title{get_variable_id}
\usage{
get_variable_id(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A character vector
}
\description{
Get sample names
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
