% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{get_variable_number}
\alias{get_variable_number}
\title{get_variable_number}
\usage{
get_variable_number(object)
}
\arguments{
\item{object}{tidymass class object.}
}
\value{
A numeric.
}
\description{
Number of variables
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
