% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{head.mass_dataset}
\alias{head.mass_dataset}
\title{head}
\usage{
\method{head}{mass_dataset}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{Additional parameters.}
}
\value{
mass_dataset class object
}
\description{
head
}
