% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_plot.R
\docType{methods}
\name{intensity_plot}
\alias{intensity_plot}
\title{intensity_plot}
\usage{
intensity_plot(
  object,
  variable_id,
  variable_index,
  color_by,
  order_by,
  desc = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{object}{(required) A mass_dataset object.}

\item{variable_id}{Variable id.}

\item{variable_index}{Variable index. You need to provide this or variable_id.}

\item{color_by}{which column (in variable_info) is used to color variables}

\item{order_by}{which column (in variable_info) is used to order variables}

\item{desc}{Descend the order? TRUE or FALSE.}

\item{interactive}{Interactive plot or not. TRUE or FALSE.}
}
\value{
A ggplot2 class object.
}
\description{
Get intensity distributation plot for one specific feature.
}
\examples{
data("expression_data")
data("sample_info")
data("sample_info_note")
data("variable_info")
data("variable_info_note")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
    sample_info_note = sample_info_note,
    variable_info_note = variable_info_note
  )
object

intensity_plot(object = object, variable_index = 1)
intensity_plot(object = object, variable_index = 1, color_by = "class")
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
