% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R, R/mass_dataset-methods.R
\name{sample_info,mass_dataset-method}
\alias{sample_info,mass_dataset-method}
\alias{expression_data,mass_dataset-method}
\alias{variable_info,mass_dataset-method}
\alias{process_info,mass_dataset-method}
\alias{ms2_data,mass_dataset-method}
\alias{show,mass_dataset-method}
\alias{dim.mass_dataset}
\alias{nrow.mass_dataset}
\alias{ncol.mass_dataset}
\alias{colnames,mass_dataset-method}
\alias{rownames,mass_dataset-method}
\alias{[.mass_dataset}
\title{colnames.mass_dataset <- function(x)}
\usage{
\S4method{sample_info}{mass_dataset}(object)

\S4method{expression_data}{mass_dataset}(object)

\S4method{variable_info}{mass_dataset}(object)

\S4method{process_info}{mass_dataset}(object)

\S4method{ms2_data}{mass_dataset}(object)

\S4method{show}{mass_dataset}(object)

\method{dim}{mass_dataset}(x)

\method{nrow}{mass_dataset}(x)

\method{ncol}{mass_dataset}(x)

\S4method{colnames}{mass_dataset}(x)

\S4method{rownames}{mass_dataset}(x)

\method{[}{mass_dataset}(x, i, j, drop = FALSE, ...)
}
\value{
sample_info

expression_data

variable_info

process_info

ms2_data

message

message

message

message

message

message

mass_dataset
}
\description{
colnames.mass_dataset <- function(x)
}
