% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{class}
\name{mass_dataset-class}
\alias{mass_dataset-class}
\title{An S4 class that stores the MS dataset}
\description{
An S4 class that stores the MS dataset
}
\section{Slots}{

\describe{
\item{\code{expression_data}}{expression data}

\item{\code{ms2_data}}{ms2 data}

\item{\code{annotation_table}}{annotation table}

\item{\code{sample_info}}{sample info}

\item{\code{variable_info}}{variable info}

\item{\code{sample_info_note}}{sample info note}

\item{\code{variable_info_note}}{variable info note}

\item{\code{process_info}}{process info}

\item{\code{other_files}}{Other files}

\item{\code{version}}{version}

\item{\code{activated}}{activated}
}}

