% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extracting.R, R/mass_dataset-methods.R,
%   R/mass_dataset_class.R, R/tinytools_mz_rt_match.R
\docType{class}
\name{sample_info,mass_dataset-method}
\alias{sample_info,mass_dataset-method}
\alias{expression_data,mass_dataset-method}
\alias{variable_info,mass_dataset-method}
\alias{process_info,mass_dataset-method}
\alias{ms2_data,mass_dataset-method}
\alias{sample_info_note,mass_dataset-method}
\alias{variable_info_note,mass_dataset-method}
\alias{dim.mass_dataset}
\alias{nrow.mass_dataset}
\alias{ncol.mass_dataset}
\alias{colnames,mass_dataset-method}
\alias{rownames,mass_dataset-method}
\alias{[.mass_dataset}
\alias{apply,mass_dataset-method}
\alias{log,mass_dataset-method}
\alias{abs,mass_dataset-method}
\alias{sqrt,mass_dataset-method}
\alias{scale,mass_dataset-method}
\alias{intersect,mass_dataset-method}
\alias{cbind.mass_dataset}
\alias{rbind.mass_dataset}
\alias{+,mass_dataset,numeric-method}
\alias{-,mass_dataset,numeric-method}
\alias{*,mass_dataset,numeric-method}
\alias{/,mass_dataset,numeric-method}
\alias{colSums,mass_dataset-method}
\alias{rowSums,mass_dataset-method}
\alias{colMeans,mass_dataset-method}
\alias{rowMeans,mass_dataset-method}
\alias{length,mass_dataset-method}
\alias{names,mass_dataset-method}
\alias{dimnames,mass_dataset-method}
\alias{head,mass_dataset-method}
\alias{tail,mass_dataset-method}
\alias{mass_dataset-class}
\alias{mz_rt_match,mass_dataset-method}
\title{An S4 class that stores the MS dataset}
\usage{
\S4method{sample_info}{mass_dataset}(object)

\S4method{expression_data}{mass_dataset}(object)

\S4method{variable_info}{mass_dataset}(object)

\S4method{process_info}{mass_dataset}(object)

\S4method{ms2_data}{mass_dataset}(object)

\S4method{sample_info_note}{mass_dataset}(object)

\S4method{variable_info_note}{mass_dataset}(object)

\method{dim}{mass_dataset}(x)

\method{nrow}{mass_dataset}(x)

\method{ncol}{mass_dataset}(x)

\S4method{colnames}{mass_dataset}(x)

\S4method{rownames}{mass_dataset}(x)

\method{[}{mass_dataset}(x, i, j, drop = FALSE, ...)

\S4method{apply}{mass_dataset}(X, MARGIN, FUN, ..., simplify = TRUE)

\S4method{log}{mass_dataset}(x, base = exp(1))

\S4method{abs}{mass_dataset}(x)

\S4method{sqrt}{mass_dataset}(x)

\S4method{scale}{mass_dataset}(x, center = TRUE, scale = TRUE)

\S4method{intersect}{mass_dataset}(x, y)

\method{cbind}{mass_dataset}(..., deparse.level = 1)

\method{rbind}{mass_dataset}(..., deparse.level = 1)

\S4method{+}{mass_dataset,numeric}(e1, e2)

\S4method{-}{mass_dataset,numeric}(e1, e2)

\S4method{*}{mass_dataset,numeric}(e1, e2)

\S4method{/}{mass_dataset,numeric}(e1, e2)

\S4method{colSums}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{rowSums}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{colMeans}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{rowMeans}{mass_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{length}{mass_dataset}(x)

\S4method{names}{mass_dataset}(x)

\S4method{dimnames}{mass_dataset}(x)

\S4method{head}{mass_dataset}(x, ...)

\S4method{tail}{mass_dataset}(x, ...)

\S4method{mz_rt_match}{mass_dataset}(
  data1,
  data2,
  mz.tol = 10,
  rt.tol = 30,
  rt.error.type = c("abs", "relative")
)
}
\arguments{
\item{object}{mass_dataset class object}

\item{x}{x}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{...}{mass_datasets}

\item{X}{X}

\item{MARGIN}{MARGIN}

\item{FUN}{FUN}

\item{simplify}{simplify}

\item{base}{base}

\item{center}{center}

\item{scale}{scale}

\item{y}{y}

\item{deparse.level}{deparse.level}

\item{e1}{a mass_dataset class object}

\item{e2}{numeric}

\item{na.rm}{na.rm}

\item{dims}{dims}

\item{data1}{data1}

\item{data2}{data2}

\item{mz.tol}{mz.tol}

\item{rt.tol}{rt.tol}

\item{rt.error.type}{rt.error.type abs or relative}

\item{..}{..}
}
\value{
sample_info

expression_data

variable_info

process_info

ms2_data

sample_info_note

variable_info_note

message

message

message

message

message

mass_dataset

result

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset

mass_dataset

mass_dataset object

mass_dataset object

mass_dataset object

mass_dataset object

vector object

vector object

vector object

vector object

vector object

vector object

vector object

mass_dataset class object

mass_dataset class object

mass_dataset object
}
\description{
An S4 class that stores the MS dataset
}
\section{Slots}{

\describe{
\item{\code{expression_data}}{expression_data}

\item{\code{ms2_data}}{ms2_data}

\item{\code{annotation_table}}{annotation_table}

\item{\code{sample_info}}{sample_info}

\item{\code{variable_info}}{variable_info}

\item{\code{sample_info_note}}{sample_info_note}

\item{\code{variable_info_note}}{variable_info_note}

\item{\code{process_info}}{process_info}

\item{\code{version}}{version}

\item{\code{activated}}{activated}
}}

