% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R, R/tinytools_mz_rt_match.R
\docType{class}
\name{mass_dataset-class}
\alias{mass_dataset-class}
\alias{mz_rt_match.mass_dataset}
\title{An S4 class that stores the MS dataset}
\usage{
\method{mz_rt_match}{mass_dataset}(
  data1,
  data2,
  mz.tol = 10,
  rt.tol = 30,
  rt.error.type = c("abs", "relative")
)
}
\arguments{
\item{data1}{A mass_dataset object.}

\item{data2}{A mass_dataset object.}

\item{mz.tol}{mz tolerance. default is 10 ppm.}

\item{rt.tol}{rt tolerance. default is 30 seconds.}

\item{rt.error.type}{RT match tolerance. abs (absolute) or relative.}
}
\value{
A merged mass_dataset object
}
\description{
An S4 class that stores the MS dataset
}
\section{Slots}{

\describe{
\item{\code{expression_data}}{expression data}

\item{\code{ms2_data}}{ms2 data}

\item{\code{annotation_table}}{annotation table}

\item{\code{sample_info}}{sample info}

\item{\code{variable_info}}{variable info}

\item{\code{sample_info_note}}{sample info note}

\item{\code{variable_info_note}}{variable info note}

\item{\code{process_info}}{process info}

\item{\code{version}}{version}

\item{\code{activated}}{activated}
}}

