% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-massdataset_logo.R
\name{massdataset_logo}
\alias{massdataset_logo}
\title{Display the MassDataset Logo}
\usage{
massdataset_logo()
}
\value{
None. This function is called for its side effect of printing the logo and messages.
}
\description{
This function displays the logo and version information for the massdataset package.
}
\examples{
massdataset_logo()
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
