% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdataset_logo.R
\docType{methods}
\name{massdataset_logo}
\alias{massdataset_logo}
\title{massdataset_logo}
\usage{
massdataset_logo()
}
\value{
logo
}
\description{
Get the detailed information of massdataset package.
}
\examples{
massdataset_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
