% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{massdataset_packages}
\alias{massdataset_packages}
\title{List all packages in the massdataset}
\usage{
massdataset_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include massdataset in the list?}
}
\value{
massdataset packages
}
\description{
List all packages in the massdataset
}
\examples{
massdataset_packages()
}
