% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_data-methods.R, R/ms2_data_class.R
\docType{methods}
\name{write_ms2_data}
\alias{write_ms2_data}
\alias{ms2_data-class}
\title{write_ms2_data}
\usage{
write_ms2_data(
  object,
  file_type = c("mgf", "msp"),
  file_name = "ms2_data",
  path = "."
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{file_type}{(required) mgf or msp}

\item{file_name}{(required) output file name.}

\item{path}{(required) working directory.}
}
\value{
msp or mgf
}
\description{
Write ms2_data class object to msp or mgf
}
\section{Slots}{

\describe{
\item{\code{column}}{column}

\item{\code{polarity}}{polarity}

\item{\code{variable_id}}{variable_id}

\item{\code{ms2_spectrum_id}}{ms2_spectrum_id}

\item{\code{ms2_mz}}{ms2_mz}

\item{\code{ms2_rt}}{ms2_rt}

\item{\code{ms2_file}}{ms2_file}

\item{\code{ms2_spectra}}{ms2_spectra}

\item{\code{mz_tol}}{mz_tol}

\item{\code{rt_tol}}{rt_tol}
}}

\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
