% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_data_class.R
\docType{class}
\name{ms2_data-class}
\alias{ms2_data-class}
\title{An S4 class that stores the ms2_data}
\description{
An S4 class that stores the ms2_data
}
\section{Slots}{

\describe{
\item{\code{column}}{column}

\item{\code{polarity}}{polarity}

\item{\code{variable_id}}{variable_id}

\item{\code{ms2_spectrum_id}}{ms2_spectrum_id}

\item{\code{ms2_mz}}{ms2_mz}

\item{\code{ms2_rt}}{ms2_rt}

\item{\code{ms2_file}}{ms2_file}

\item{\code{ms2_spectra}}{ms2_spectra}

\item{\code{mz_tol}}{mz_tol}

\item{\code{rt_tol}}{rt_tol}
}}

