% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSet-class.R
\docType{class}
\name{mSet-class}
\alias{mSet-class}
\title{An S4 class that stores the MS dataset from MetaboAnalystR (Prof. Jeff Xia)}
\description{
An S4 class that stores the MS dataset from MetaboAnalystR (Prof. Jeff Xia)
}
\section{Slots}{

\describe{
\item{\code{rawfiles}}{character}

\item{\code{rawOnDisk}}{rawOnDisk}

\item{\code{rawInMemory}}{rawInMemory}

\item{\code{params}}{params}

\item{\code{peakpicking}}{peakpicking}

\item{\code{peakgrouping}}{peakgrouping}

\item{\code{peakRTcorrection}}{peakRTcorrection}

\item{\code{peakfilling}}{peakfilling}

\item{\code{peakAnnotation}}{peakAnnotation}

\item{\code{dataSet}}{dataSet}

\item{\code{runningplan}}{runningplan}

\item{\code{msgSet}}{msgSet}

\item{\code{userpath}}{userpath}

\item{\code{WorkingDir}}{WorkingDir}
}}

