% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_ms2.R
\name{mutate_ms2}
\alias{mutate_ms2}
\title{Mutate MS2 spectra for each variable}
\usage{
mutate_ms2(
  object,
  column = c("rp", "hilic"),
  polarity = c("positive", "negative"),
  ms1.ms2.match.mz.tol = 15,
  ms1.ms2.match.rt.tol = 30,
  path = "."
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{column}{rp or hilic.}

\item{polarity}{positive or negative}

\item{ms1.ms2.match.mz.tol}{default is 15 ppm.}

\item{ms1.ms2.match.rt.tol}{default is 30 seconds.}

\item{path}{work directory.}
}
\value{
A mass_dataset object.
}
\description{
Mutate MS2 spectra for each variable
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
library(massdataset)

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
