% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_rsd.R
\name{mutate_rsd}
\alias{mutate_rsd}
\title{Mutate Relative Standard Deviation (RSD) in mass_dataset Object}
\usage{
mutate_rsd(object, according_to_samples = "all")
}
\arguments{
\item{object}{A mass_dataset object.}

\item{according_to_samples}{A character vector specifying the sample IDs to consider when
calculating the RSD. Default is "all", which considers all samples.}
}
\value{
A modified mass_dataset object with an updated \code{variable_info} slot.
}
\description{
This function adds a new column to the \code{variable_info} slot of a mass_dataset object,
which contains the relative standard deviation (RSD) for each variable according to the samples specified.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##calculate RSDs according to all the samples
object =
  mutate_rsd(object = object)

object

head(extract_variable_info(object))

##calculate RSDs according to only QC samples
object =
  mutate_rsd(object = object,
                according_to_samples =
              get_sample_id(object)[extract_sample_info(object)$class == "QC"])

object

head(extract_variable_info(object))
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
