% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sample_na_number.R
\name{mutate_sample_na_freq}
\alias{mutate_sample_na_freq}
\title{Add NA frequency for each sample to sample_info}
\usage{
mutate_sample_na_freq(object, according_to_variables = "all")
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{according_to_variables}{(required) What variables used to filter variables.
Default is "all". If you
want to use only several variables, provide their names as a vector.}
}
\value{
A mass_dataset class object
}
\description{
Add NA frequency for each sample to sample_info
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object

##calculate NA frequency according to all the variables
object2 =
  mutate_sample_na_freq(object = object)

head(extract_sample_info(object))
head(extract_sample_info(object2))

##calculate NA frequency according to only variables with mz > 100
variable_id =
object2 \%>\%
  activate_mass_dataset(what = "variable_info") \%>\%
  filter(mz > 100) \%>\%
  pull(variable_id)

object3 =
  mutate_sample_na_freq(object = object2,
                according_to_variables = variable_id)

object3

head(extract_sample_info(object3))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
