% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variable_na_number.R
\name{mutate_variable_na_number}
\alias{mutate_variable_na_number}
\title{Add NA number for each feature to variable_info}
\usage{
mutate_variable_na_number(object, according_to_samples = "all")
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{according_to_samples}{(required) What samples used to filter variables.
Default is "all". If you
want to use only several samples, provide their names as a vector.}
}
\value{
A mass_dataset class object
}
\description{
Add NA number for each feature to variable_info
}
\examples{
library(massdataset)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##calculate NA number according to all the samples
object2 =
  mutate_variable_na_number(object = object)

head(extract_variable_info(object))
head(extract_variable_info(object2))

##calculate NA number according to only QC samples
object3 =
  mutate_variable_na_number(object = object2,
                according_to_samples =
              get_sample_id(object)[extract_sample_info(object)$class == "QC"])

object3

head(extract_variable_info(object3))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
