% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variable_zero.R
\name{mutate_variable_zero_freq}
\alias{mutate_variable_zero_freq}
\title{Mutate Variable Zero Frequency in mass_dataset Object}
\usage{
mutate_variable_zero_freq(object, according_to_samples = "all")
}
\arguments{
\item{object}{A mass_dataset object.}

\item{according_to_samples}{A character vector specifying the sample IDs to consider when
calculating zero frequency. Default is "all", which considers all samples.}
}
\value{
A modified mass_dataset object with an updated \code{variable_info} slot.
}
\description{
This function adds a new column to the \code{variable_info} slot of a mass_dataset object,
which contains the frequency of zeros for each variable according to the samples specified.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##calculate NA frequency according to all the samples
object2 =
  mutate_variable_zero_freq(object = object)

head(extract_variable_info(object))
head(extract_variable_info(object2))

##calculate NA number according to only QC samples
object3 =
  mutate_variable_zero_freq(object = object2,
                according_to_samples =
              get_sample_id(object)[extract_sample_info(object)$class == "QC"])

object3

head(extract_variable_info(object3))
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
