% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_variable_zero.R
\name{mutate_variable_zero_number}
\alias{mutate_variable_zero_number}
\title{Mutate Variable Zero Number in mass_dataset Object}
\usage{
mutate_variable_zero_number(object, according_to_samples = "all")
}
\arguments{
\item{object}{A mass_dataset object.}

\item{according_to_samples}{A character vector specifying the sample IDs to consider when
counting zeros. Default is "all", which considers all samples.}
}
\value{
A modified mass_dataset object with an updated \code{variable_info} slot.
}
\description{
This function adds a new column to the \code{variable_info} slot of a mass_dataset object,
which contains the number of zeros for each variable according to the samples specified.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##calculate NA number according to all the samples
object2 =
  mutate_variable_zero_number(object = object)

colnames(extract_variable_info(object))
colnames(extract_variable_info(object2))
object2@variable_info_note

##calculate NA number according to only QC samples
object3 <-
  mutate_variable_zero_number(object = object2,
                according_to_samples =
              get_sample_id(object)[extract_sample_info(object)$class == "QC"])

object3

colnames(extract_variable_info(object3))
object3@variable_info_note
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
