% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_in_package.R
\docType{data}
\name{mzmine_table}
\alias{mzmine_table}
\title{zmine feature table}
\format{
A data frame.
}
\description{
mzmine_table
An example feature table from mzMine.
}
\examples{
data(mzmine_table)
}
\keyword{data}
