% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_parameters.R
\docType{methods}
\name{parse_tidymass_parameter}
\alias{parse_tidymass_parameter}
\title{Parse tidymass_parameter Object to Data Frame}
\usage{
parse_tidymass_parameter(object)
}
\arguments{
\item{object}{A \code{tidymass_parameter} object to be parsed.}
}
\value{
A data frame containing the package name, function name, parameters, and the time when the function was called.
}
\description{
This function takes a \code{tidymass_parameter} object and converts it into a data frame for easier manipulation and reporting.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
object
parse_tidymass_parameter(object@process_info$create_mass_dataset)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
