% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{rbind.mass_dataset}
\alias{rbind.mass_dataset}
\title{rbind}
\usage{
\method{rbind}{mass_dataset}(..., deparse.level = 1)
}
\arguments{
\item{...}{One or more \code{mass_dataset} objects to be row bound.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. Default is 1.}
}
\value{
mass_dataset
}
\description{
rbind
}
