% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mztab2mass_dataset.R
\name{read_mztab}
\alias{read_mztab}
\title{Read mzTab-M format data}
\usage{
read_mztab(file, path = ".")
}
\arguments{
\item{file}{File name of the mzTab-M data.}

\item{path}{Where is the file? Default is .}
}
\value{
A list.
}
\description{
Read mzTab-M format data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
