% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_arrange.R, R/dplyr_filter.R,
%   R/dplyr_left_join.R, R/dplyr_mutate.R, R/dplyr_rename.R, R/dplyr_select.R,
%   R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{desc}
\alias{filter}
\alias{left_join}
\alias{mutate}
\alias{transmute}
\alias{mutate_all}
\alias{mutate_at}
\alias{n}
\alias{rename}
\alias{select}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:mutate]{transmute}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

