% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_parameters.R
\name{report_parameters}
\alias{report_parameters}
\title{report_parameters}
\usage{
report_parameters(object, path = ".")
}
\arguments{
\item{object}{(required) tidymass_parameters class object.}

\item{path}{working directory}
}
\value{
html report
}
\description{
report_parameters
}
\examples{
\dontrun{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
object

library(tidyverse)

object =
object \%>\%
  activate_mass_dataset(what = "expression_data") \%>\%
  filter(!is.na(QC_1))

object =
  object \%>\%
  activate_mass_dataset(what = "expression_data") \%>\%
  filter(!is.na(QC_2))


report_parameters(object = object, path = "demo_data")
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
