% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_parameters.R
\docType{methods}
\name{report_parameters}
\alias{report_parameters}
\title{Generate Parameter Report for mass_dataset Object}
\usage{
report_parameters(object, path = ".")
}
\arguments{
\item{object}{A \code{mass_dataset} object for which the parameter report is to be generated.}

\item{path}{The directory where the parameter report will be saved. Default is the current working directory.}
}
\value{
An HTML report summarizing the parameters used in the processing of the \code{mass_dataset} object. The report is saved in the specified directory.
}
\description{
This function generates an HTML report summarizing the parameters used in the processing of a \code{mass_dataset} object. The report is saved in a specified directory.
}
\examples{
\dontrun{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
object

library(tidyverse)

object =
object \%>\%
  activate_mass_dataset(what = "expression_data") \%>\%
  filter(!is.na(QC_1))

object =
  object \%>\%
  activate_mass_dataset(what = "expression_data") \%>\%
  filter(!is.na(QC_2))

object =
object \%>\%
  mutate_mean_intensity()

object =
  object \%>\%
  mutate_median_intensity() \%>\%
  mutate_rsd()

report_parameters(object = object, path = "demo_data")
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
