% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R
\docType{methods}
\name{rownames,mass_dataset-method}
\alias{rownames,mass_dataset-method}
\alias{rownames}
\title{rownames}
\usage{
\S4method{rownames}{mass_dataset}(x)
}
\arguments{
\item{x}{object}
}
\description{
rownames
}
