% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R, R/5-data_in_package.R
\docType{data}
\name{sample_info}
\alias{sample_info}
\title{Generic Function for Retrieving Sample Information}
\format{
A data frame with 8 rows (8 samples) and 4 columns.
}
\usage{
sample_info(object, ...)
}
\arguments{
\item{object}{An object from which sample information will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation.
}
\description{
The \code{sample_info} generic function retrieves sample-related information from a provided object.

sample_info
A data frame contains the sample information.
}
\examples{
data(sample_info)
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
\keyword{data}
