% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\docType{data}
\name{sample_info}
\alias{sample_info}
\title{Sample information}
\format{
A data frame with 8 rows (8 samples) and 4 columns.
}
\usage{
sample_info
}
\description{
sample_info
A data frame contains the sample information.
}
\keyword{datasets}
