% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_info_note.R
\docType{data}
\name{sample_info_note}
\alias{sample_info_note}
\title{Sample information note}
\format{
A data frame with 4 rows (4 columns in sample_info) and 2 columns.
}
\usage{
sample_info_note
}
\description{
sample_info_note
A data frame contains description for each sample information column.
}
\keyword{datasets}
