% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_dataset_class.R,
%   R/tidymass_parameter_class.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
\S4method{show}{mass_dataset}(object)

\S4method{show}{tidymass_parameter}(object)
}
\arguments{
\item{object}{A \code{tidymass_parameter} instance.}
}
\value{
message

message
}
\description{
show method for mass_dataset class

show method for tidymass_parameter
}
\author{
Xiaotao Shen

Xiaotao Shen
}
