% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_explore_plot.R
\name{show_missing_values}
\alias{show_missing_values}
\title{Add NA number for each feature to variable_info}
\usage{
show_missing_values(
  object,
  show_row_names = FALSE,
  show_column_names = TRUE,
  column_names_rot,
  cell_color = "transparent",
  row_names_side = "right",
  percentage = FALSE,
  ...
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{show_row_names}{show_row_names. see?ComplexHeatmap::Heatmap}

\item{show_column_names}{show_column_names see?ComplexHeatmap::Heatmap}

\item{column_names_rot}{column_names_rot see?ComplexHeatmap::Heatmap}

\item{cell_color}{cell color}

\item{row_names_side}{row names side}

\item{percentage}{percentage or not}

\item{...}{Other parameters for ComplexHeatmap::Heatmap}
}
\value{
A ggplot2 object
}
\description{
Add NA number for each feature to variable_info
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##show missing values plot
show_missing_values(object)

show_missing_values(object[1:10,], cell_color = "white")

###only show subject samples
object \%>\%
  activate_mass_dataset(what = "sample_info") \%>\%
  filter(class == "Subject") \%>\%
  show_missing_values()

###only show QC samples
object \%>\%
  activate_mass_dataset(what = "expression_data") \%>\%
  dplyr::select(contains("QC")) \%>\%
  show_missing_values()

###only show features with mz < 100
object \%>\%
  activate_mass_dataset(what = "variable_info") \%>\%
  dplyr::filter(mz < 100) \%>\%
  show_missing_values(cell_color = "white",
                      show_row_names = TRUE,
                      row_names_side = "left", 
                      percentage = TRUE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
