% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_mz_rt_plot.R
\name{show_mz_rt_plot}
\alias{show_mz_rt_plot}
\title{mz and rt plot}
\usage{
show_mz_rt_plot(object, hex = FALSE)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{hex}{hex or not.}
}
\value{
A ggplot2 object
}
\description{
mz and rt plot
}
\examples{
library(massdataset)
library(plyr)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object \%>\%
  show_mz_rt_plot()

object \%>\%
  log(10) \%>\%
  show_mz_rt_plot()

object \%>\%
  show_mz_rt_plot(hex = TRUE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
