% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_explore_plot.R
\name{show_sample_missing_values}
\alias{show_sample_missing_values}
\title{show_sample_missing_values}
\usage{
show_sample_missing_values(
  object,
  color_by,
  order_by,
  percentage = FALSE,
  desc = FALSE
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{color_by}{which column (sample_info) is used to color samples}

\item{order_by}{which column (sample_info) is used to order samples}

\item{percentage}{show MV percentage?}

\item{desc}{desc}
}
\value{
A ggplot2 object
}
\description{
show_sample_missing_values
}
\examples{
library(massdataset)
library(plyr)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##show missing values plot
show_sample_missing_values(object)
show_sample_missing_values(object, color_by = "class")
show_sample_missing_values(object, color_by = "class", order_by = "na")
show_sample_missing_values(object, color_by = "class", order_by = "na",
                           desc = TRUE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
