% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_explore_plot.R
\name{show_sample_missing_values}
\alias{show_sample_missing_values}
\title{Add NA number for each feature to variable_info}
\usage{
show_sample_missing_values(
  object,
  color_by,
  order_by,
  frequency = FALSE,
  desc = FALSE
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{show_row_names}{show_row_names. see?ComplexHeatmap::Heatmap}

\item{show_column_names}{show_column_names see?ComplexHeatmap::Heatmap}

\item{column_names_rot}{column_names_rot see?ComplexHeatmap::Heatmap}

\item{cell_color}{cell color}

\item{row_names_side}{row names side}

\item{...}{Other parameters for ComplexHeatmap::Heatmap}
}
\value{
A ggplot2 object
}
\description{
Add NA number for each feature to variable_info
}
\examples{
library(massdataset)
library(plyr)
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##show missing values plot
show_sample_missing_values(object)
show_sample_missing_values(object, color_by = "class")
show_sample_missing_values(object, color_by = "class", order_by = "na")
show_sample_missing_values(object, color_by = "class", order_by = "na",
                           desc = TRUE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
