% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_explore_plot.R
\name{show_variable_missing_values}
\alias{show_variable_missing_values}
\title{show_variable_missing_values}
\usage{
show_variable_missing_values(
  object,
  color_by,
  order_by,
  frequency = FALSE,
  show_x_text = FALSE,
  show_x_ticks = FALSE,
  desc = FALSE
)
}
\arguments{
\item{object}{(required) mass_dataset class object.}

\item{color_by}{which column (variable_info) is used to color variables}

\item{order_by}{which column (variable_info) is used to order variables}

\item{frequency}{show MV frequency?}

\item{show_x_text}{show_x_text}

\item{show_x_ticks}{show_x_ticks}

\item{desc}{desc}
}
\value{
A ggplot2 object
}
\description{
show_variable_missing_values
}
\examples{
library(massdataset)
library(plyr)
library(tidyverse)
data("expression_data")
data("variable_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

##show missing values plot
show_variable_missing_values(object)
show_variable_missing_values(object, color_by = "mz")

show_variable_missing_values(object, color_by = "rt") +
  scale_color_gradient(low = "skyblue", high = "red") 

show_variable_missing_values(object, color_by = "mz", 
                             order_by = "na")
show_variable_missing_values(object, color_by = "mz", 
                             order_by = "na",
                           desc = TRUE, frequency = TRUE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
