% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-mass_dataset.R
\docType{methods}
\name{split_mass_dataset}
\alias{split_mass_dataset}
\title{Split and Update mass_dataset Object Based on Criteria}
\usage{
split_mass_dataset(object, by, fun)
}
\arguments{
\item{object}{A \code{mass_dataset} object to be split.}

\item{by}{The column name in \code{sample_info} or \code{variable_info} used for splitting.}

\item{fun}{A function to apply when \code{by} is numeric. This function should return a logical vector.}
}
\value{
A list of subsetted \code{mass_dataset} objects.
}
\description{
This function splits a \code{mass_dataset} object into multiple
subsets based on a specified column in either \code{sample_info}
or \code{variable_info}. It also updates the \code{process_info} slot with
the splitting parameters.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object <-
activate_mass_dataset(object, what = "sample_info")

new_object <-
  split_mass_dataset(object = object, by = "group")

new_object \%>\% lapply(dim)
new_object \%>\% lapply(colnames)

object <-
  activate_mass_dataset(object, what = "variable_info")

new_object <-
  split_mass_dataset(object = object, by = "rt", fun = function(rt) rt > 600)

new_object \%>\% lapply(dim)
plot(extract_variable_info(new_object[[1]])$rt)
plot(extract_variable_info(new_object[[2]])$rt)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
