% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-mass_dataset.R
\name{[.mass_dataset}
\alias{[.mass_dataset}
\title{[ method}
\usage{
\method{[}{mass_dataset}(x, i, j, drop = FALSE, ...)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{j}{j}

\item{drop}{drop}

\item{...}{Other parameters}
}
\value{
mass_dataset
}
\description{
Subsetting method for mass_dataset using [ operator. This allows extracting rows (variables) and columns (samples) from the object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
