% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summarize_samples.R
\name{summarise_samples}
\alias{summarise_samples}
\alias{summarize_samples}
\alias{summarise_samples.mass_dataset}
\title{merge samples}
\usage{
summarise_samples(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  group_by,
  ...
)

summarize_samples(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  group_by,
  ...
)

\method{summarise_samples}{mass_dataset}(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  group_by,
  ...
)
}
\arguments{
\item{object}{mass_dataset}

\item{what}{which you want to use}

\item{group_by}{summarize samples by}

\item{...}{other params}
}
\value{
mass_dataset
}
\description{
merge samples
}
