% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-mass_dataset.R
\docType{methods}
\name{get_sample_number}
\alias{get_sample_number}
\alias{get_variable_number}
\alias{get_sample_id}
\alias{get_variable_id}
\alias{get_mv_number}
\alias{dim.mass_dataset}
\alias{nrow.mass_dataset}
\alias{ncol.mass_dataset}
\alias{colnames.mass_dataset}
\alias{rownames.mass_dataset}
\alias{apply,mass_dataset-method}
\alias{intersect,mass_dataset,mass_dataset-method}
\alias{summary.mass_dataset}
\alias{length.mass_dataset}
\alias{names.mass_dataset}
\alias{dimnames.mass_dataset}
\alias{is.na.mass_dataset}
\title{get_sample_number}
\usage{
get_sample_number(object)

get_variable_number(object)

get_sample_id(object)

get_variable_id(object)

get_mv_number(
  object,
  by = c("total", "sample", "variable"),
  show_by = c("number", "percentage")
)

\method{dim}{mass_dataset}(x)

\method{nrow}{mass_dataset}(x)

\method{ncol}{mass_dataset}(x)

\method{colnames}{mass_dataset}(x)

\method{rownames}{mass_dataset}(x)

\S4method{apply}{mass_dataset}(X, MARGIN, FUN, ..., simplify = TRUE)

\S4method{intersect}{mass_dataset,mass_dataset}(x, y)

\method{summary}{mass_dataset}(object, ...)

\method{length}{mass_dataset}(x)

\method{names}{mass_dataset}(x)

\method{dimnames}{mass_dataset}(x)

\method{is.na}{mass_dataset}(x)
}
\arguments{
\item{object}{object}

\item{by}{total: Missing value number in total. sample: Missing value number in each sample.
variable: Missing value number in each variable.}

\item{show_by}{number: missing value number. percentage: missing value percentage.}

\item{x}{x}

\item{X}{X}

\item{MARGIN}{MARGIN}

\item{FUN}{FUN}

\item{...}{other parameters}

\item{simplify}{simplify}

\item{y}{y}
}
\value{
A numeric.

A numeric.

A character vector.

A character vector.

A numeric (vector).

message

message

message

message

message

result

result

vector object

vector object

vector object

message
}
\description{
Number of samples

Number of variables

Sample IDs

Variable IDs

Get missing value number/percentage in expression data.

###old version
#' @title apply
#' @method apply mass_dataset
#' @param X X
#' @param MARGIN MARGIN
#' @param FUN FUN
#' @param ... ...
#' @param simplify simplify
#' @export
#' @rdname summary-mass_dataset
#' @return result

#' @title intersect
#' @method intersect mass_dataset
#' @param x x
#' @param y y
#' @export
#' @rdname summary-mass_dataset
#' @return result
}
\details{
apply.mass_dataset =
function(X, MARGIN, FUN, ..., simplify = TRUE){
apply(as.matrix(X@expression_data), MARGIN, FUN, ..., simplify = simplify)
}

intersect.mass_dataset = function(x, y){
intersect(x@sample_info$sample_id,
y@sample_info$sample_id)
}
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_sample_number(object = object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_variable_number(object = object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_sample_id(object = object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 head(get_variable_id(object = object))
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
object
head(get_variable_id(object = object))
get_mv_number(object)
get_mv_number(object, by = "sample")
head(get_mv_number(object, by = "variable", "percentage"))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
