% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masstools_mz_rt_match.R,
%   R/summary-mass_dataset.R
\docType{methods}
\name{mz_rt_match,mass_dataset,mass_dataset-method}
\alias{mz_rt_match,mass_dataset,mass_dataset-method}
\alias{get_sample_number}
\alias{get_variable_number}
\alias{get_sample_id}
\alias{get_variable_id}
\alias{get_mv_number}
\alias{dim.mass_dataset}
\alias{nrow.mass_dataset}
\alias{ncol.mass_dataset}
\alias{colnames.mass_dataset}
\alias{rownames.mass_dataset}
\alias{apply,mass_dataset-method}
\alias{intersect.mass_dataset}
\alias{summary.mass_dataset}
\alias{length.mass_dataset}
\alias{names.mass_dataset}
\alias{dimnames.mass_dataset}
\title{mz_rt_match}
\usage{
\S4method{mz_rt_match}{mass_dataset,mass_dataset}(
  data1,
  data2,
  mz.tol = 10,
  rt.tol = 30,
  rt.error.type = c("abs", "relative")
)

get_sample_number(object)

get_variable_number(object)

get_sample_id(object)

get_variable_id(object)

get_mv_number(
  object,
  by = c("total", "sample", "variable"),
  show_by = c("number", "percentage")
)

\method{dim}{mass_dataset}(x)

\method{nrow}{mass_dataset}(x)

\method{ncol}{mass_dataset}(x)

\method{colnames}{mass_dataset}(x)

\method{rownames}{mass_dataset}(x)

\S4method{apply}{mass_dataset}(X, MARGIN, FUN, ..., simplify = TRUE)

\method{intersect}{mass_dataset}(x, y)

\method{summary}{mass_dataset}(object, ...)

\method{length}{mass_dataset}(x)

\method{names}{mass_dataset}(x)

\method{dimnames}{mass_dataset}(x)
}
\arguments{
\item{data1}{A mass_dataset object.}

\item{data2}{A mass_dataset object.}

\item{mz.tol}{mz tolerance. default is 10 ppm.}

\item{rt.tol}{rt tolerance. default is 30 seconds.}

\item{rt.error.type}{RT match tolerance. abs (absolute) or relative.}

\item{object}{object}

\item{by}{total: Missing value number in total. sample: Missing value number in each sample.
variable: Missing value number in each variable.}

\item{show_by}{number: missing value number. percentage: missing value percentage.}

\item{x}{x}

\item{X}{X}

\item{MARGIN}{MARGIN}

\item{FUN}{FUN}

\item{...}{other parameters}

\item{simplify}{simplify}

\item{y}{y}
}
\value{
A merged mass_dataset object

A numeric.

A numeric.

A character vector.

A character vector.

A numeric (vector).

message

message

message

message

message

result

mass_dataset object

vector object

vector object

vector object
}
\description{
Number of samples

Number of variables

Sample IDs

Variable IDs

Get missing value number/percentage in expression data.

###old version
#' @title apply
#' @method apply mass_dataset
#' @param X X
#' @param MARGIN MARGIN
#' @param FUN FUN
#' @param ... ...
#' @param simplify simplify
#' @export
#' @rdname summary-mass_dataset
#' @return result
}
\details{
apply.mass_dataset =
function(X, MARGIN, FUN, ..., simplify = TRUE){
apply(as.matrix(X@expression_data), MARGIN, FUN, ..., simplify = simplify)
}
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_sample_number(object = object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_variable_number(object = object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 get_sample_id(object = object)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
 head(get_variable_id(object = object))
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )
object
head(get_variable_id(object = object))
get_mv_number(object)
get_mv_number(object, by = "sample")
head(get_mv_number(object, by = "variable", "percentage"))
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
