% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-mass_dataset.R
\name{tail.mass_dataset}
\alias{tail.mass_dataset}
\title{tail}
\usage{
\method{tail}{mass_dataset}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{Additional parameters.}
}
\value{
mass_dataset class object
}
\description{
tail
}
