% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_parameter_class.R
\docType{class}
\name{tidymass_parameter-class}
\alias{tidymass_parameter-class}
\title{tidymass_parameter S4 Class}
\description{
An S4 class to represent the parameters used in tidymass functions.
}
\section{Slots}{

\describe{
\item{\code{pacakge_name}}{A character vector representing the name of the package where the function resides.}

\item{\code{function_name}}{A character vector representing the name of the function that uses these parameters.}

\item{\code{parameter}}{A list containing the parameters used in the function.}

\item{\code{time}}{A POSIXct object representing the time when the function was called.}
}}

\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
