% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymass_parameter-methods.R
\docType{methods}
\name{translate_tidymass_parameter}
\alias{translate_tidymass_parameter}
\title{translate_tidymass_parameter}
\usage{
translate_tidymass_parameter(object)
}
\arguments{
\item{object}{(required) tidymass_parameters class object.}
}
\value{
A data.frame
}
\description{
translate tidymass_parameter to data.frame.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )
object
translate_tidymass_parameter(object@process_info[[1]])
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
