% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mass_dataset.R
\name{update_mass_dataset}
\alias{update_mass_dataset}
\title{Update a mass_dataset Object}
\usage{
update_mass_dataset(object)
}
\arguments{
\item{object}{A \code{mass_dataset} object that you want to update.}
}
\value{
A \code{mass_dataset} object with updated slots and process information.
}
\description{
This function updates the slots of a \code{mass_dataset} object to ensure consistency
among \code{sample_info}, \code{variable_info}, and \code{expression_data}. It also logs the
update process.
}
\examples{
data("expression_data")
data("sample_info")
data("variable_info")

object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info,
  )

object

####only remain QC samples
sample_info = extract_sample_info(object = object)

sample_info =
  dplyr::filter(sample_info, class == "QC")

object@sample_info = sample_info

object = update_mass_dataset(object = object)
object

####only remain feature with mz < 300
variable_info = extract_variable_info(object = object)

variable_info =
  dplyr::filter(variable_info, mz < 300)

object@variable_info = variable_info

object = update_mass_dataset(object = object)
object
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
