% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mass_dataset.R
\name{update_sample_info}
\alias{update_sample_info}
\title{Update Sample Information in a mass_dataset Object}
\usage{
update_sample_info(object)
}
\arguments{
\item{object}{A \code{mass_dataset} object whose \code{sample_info} and \code{sample_info_note} slots you want to update.}
}
\value{
A \code{mass_dataset} object with updated \code{sample_info} and \code{sample_info_note} slots.
}
\description{
This function updates the \code{sample_info} and \code{sample_info_note} slots of a \code{mass_dataset} object.
It ensures that the columns in \code{sample_info} are consistent with the names in \code{sample_info_note}.
}
\examples{
\dontrun{
# Assuming `md` is a mass_dataset object
updated_md <- update_sample_info(md)
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
