% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mass_dataset.R
\docType{methods}
\name{update_sample_info}
\alias{update_sample_info}
\title{update_sample_info}
\usage{
update_sample_info(object)
}
\arguments{
\item{object}{(required) mass_dataset class object.}
}
\value{
A mass_dataset class object.
}
\description{
update_sample_info
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
