% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_in_package.R
\docType{data}
\name{variable_info}
\alias{variable_info}
\title{Variable information}
\format{
A data frame with 1000 rows (1000 variables) and 3 columns.
}
\description{
variable_info
A data frame contains the variable information.
}
\examples{
data(variable_info)
}
\keyword{data}
