% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_in_package.R
\docType{data}
\name{variable_info_note}
\alias{variable_info_note}
\title{Variable information note}
\format{
A data frame with 3 rows (3 columns in variable_info) and 2 columns.
}
\description{
variable_info_note
A data frame contains description for each variable information column.
}
\examples{
data(variable_info_note)
}
\keyword{data}
