% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-all_generics.R, R/5-data_in_package.R
\docType{data}
\name{variable_info_note}
\alias{variable_info_note}
\title{Generic Function for Retrieving Variable Information Notes}
\format{
A data frame with 3 rows (3 columns in variable_info) and 2 columns.
}
\usage{
variable_info_note(object, ...)
}
\arguments{
\item{object}{An object from which variable information notes will be extracted.}

\item{...}{Additional arguments.}
}
\value{
Depends on the specific method implementation. Typically, it might return a character vector, list, or dataframe containing notes or metadata about the variable information.
}
\description{
The \code{variable_info_note} generic function retrieves notes or metadata associated with variable information from the provided object.

variable_info_note
A data frame contains description for each variable information column.
}
\examples{
data(variable_info_note)
}
\seealso{
\code{\link{methods}} for more details on S4 methods.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
\keyword{data}
